Initial import of linux-libertine-ttf-2.1.0:
A TrueType font-family: Serif and Grotesk (Sans Serif) for practical
use in documents. This project aims at creating a free alternative
to the standard W*ndows Fonts (T*mes and Ar*al).  PfaEdit is used
as FontEditor.  The goal is to create a really good font with a
real italic and bold variant, with OSF (old style figures -> old
numbers) and small Caps.