Fix the csound5 install phase.
At some point, pkgsrc started setting a different MAKE_ENV in the
install phase than in the build phase.  This caused scons to report that
"the build action changed", which caused scons to rebuild csound5 and
install it with a broken configuration.

This hack saves MAKE_ENV to a file in WRKDIR, then uses that file to set
the environment in the install phase.