Import eltclsh version 1.6. Based on PR 22994 by Anthony Mallet. I modify to adapt to recent pkgsrc tree and update from 1.4 to 1.6.
eltclsh (editline tcl shell) is an interactive shell for the TCL
programming language. It provides command line editing, history browsing
as well as variables and command completion thanks to editline
features. The completion engine is programmable in a way similar to tcsh,
and comes with an programmed completion for the whole TCL language by
default.

The package also provides elwish, an interactive interpreter for the Tk
toolkit.