Initial import of gamazons-0.83:
Amazons is a game played on a 10x10 chess board. Each side has four
pieces (amazons) that move like chess queens (in a straight line
in any direction). Instead of capturing pieces like in chess, the
game is determined based on who moves last.

Each move consists of two parts. First an amazon moves to a new
square and then fires an arrow to another square (the arrow is
fired in a straight line in any direction from the square the amazon
landed on). The square the arrow lands on becomes a permenant block
for the rest of the game. No one can move over it, or fire an arrow
over it. Every turn an amazon must move and fire an arrow, so every
turn there is one less square available on the board. Try and block
in your opponent or section off a good chunk of the board for
yourself.