Update txt2html to version 2.44.
The old version of txt2html was so outdated that it makes little
sense to list the changes. Most importantly, the new version now
has major revision number 2, which is due to the fact that
HTML::TextToHTML and the official txt2html were merged. Lots of
subsequent added features and bug fixes.

The updated package was prepared by Aleksey Cheusov in wip.
Closes PR pkg/35469.