Initial import of p5-Encode-Detect 1.0 from pkgsrc-wip.
The package was created by Hiramatsu Yoshifumi, with some very small changes
by me:
  - different wording in COMMENT
  - BUILD_DEPENDS instead of DEPENDS. The required version of
    p5-ExtUtils-CBuilder is not specified in Build.PL, so I left it that way.
  - Added USE_LANGUAGES.

=====================================
This Perl module is an Encode::Encoding subclass that uses
Encode::Detect::Detector to determine the charset of the input data
and then decodes it using the encoder of the detected charset.

It is similar to Encode::Guess, but does not require the configuration
of a set of expected encodings. Like Encode::Guess, it only supports
decoding--it cannot encode.