Update to 0.4.10:
gnome-speech 0.4.10:

* Fix for bug 409573 to have eSpeak driver also look for LANG when
  determining the default language. (Gilles Casse)

gnome-speech 0.4.9:

* Fix for bug 398916 to help select the Viavoice Chinese voices. Also
  fixed a memory leak with the driver. (Gilles Casse)

* Fix for bug 399512 to handle config parameters better. (Leonardo Boshell)

* Fix for bug 403763 to include math library when building driver
  for Swift. (Willie Walker)

gnome-speech 0.4.8:

* Brand new driver for eSpeak (http://espeak.sourceforge.net).
Many many thanks to Gilles Casse for his contribution here!!!

* Fix for bug 321658 to allow ibmtts is unable to select/change voice
of a different language.  Many many thanks to Gilles Casse for his
contribution here!!!

gnome-speech 0.4.7:

* Add support for Cepstral Swift driver.  There are known problems
with this support.  In particular, there are audio artifacts
(chirps and pops) when speech is interrupted.  We have a support
request into Cepstral for help with this: Support Ticket #6297.