Initial import of p5-XML-Feed 0.12.
XML::Feed is a syndication feed parser for both RSS and Atom feeds. It
also implements feed auto-discovery for finding feeds, given a URI.

The goal of XML::Feed is to provide a unified API for parsing and using
the various syndication formats. The different flavors of RSS and Atom
handle data in different ways: date handling; summaries and content;
escaping and quoting; etc. This module attempts to remove those differ-
ences by providing a wrapper around the formats and the classes imple-
menting those formats (XML::RSS and XML::Atom::Feed).