Update geeklog to 1.4.1.
pkgsrc's change: improving our README file.


Geeklog 1.4.1

New Features

  * Support for Microsoft SQL Server. Starting with this release, Geeklog can
    now also be installed on Microsoft SQL Server, so it's no longer restricted
    to just MySQL. The MS SQL support was developed by Randy Kolenko. Thanks,
    Randy!
    Please note that any third-party plugins will have to offer support for MS
    SQL before they can be installed on Microsoft SQL Server. The bundled
    plugins (Calendar, Links, Polls, Spam-X, Static Pages) have already been
    updated accordingly.
  * Calendar plugin. The formerly built-in calendar and events have now been
    moved into a separate plugin. This complements the move of the polls and
    links sections into plugins in Geeklog 1.4.0 and makes Geeklog more modular
    as you can now easily disable or replace functionality that you don't need
    for your site.
  * Multi-language support. It is now possible to build truly multi-linugal
    sites with Geeklog where not only the navigation but also the content of
    the site changes with the language.
  * Ships with FCKeditor 2.3.1, which once again includes a file manager for
    uploading images.
  * A function for mass-deletion of old or inactive users. The list
    automatically searches for users that have never logged in, only used the
    site for a very short time or have not been online since a very long time.
    The time span can be varied, and found users can be selectively deleted.

Security

In the light of the security issues discovered in Geeklog 1.4.0 and earlier
versions, the Geeklog source code has undergone a code review. We have
identified and addressed several minor issues and introduced new measures to
enhance security in this release. As a welcome side effect, the code reviews
have also uncovered a few bugs and inconsistencies that we also fixed in this
release.

Spam Protection

With this release we are finally removing support for the discontinued
MT-Blacklist. In its place, we are now using a system called Spam Link
Verification (SLV) run by Russ Jones at www.linksleeve.org. SLV could be
described as a community-driven, automatically updated blacklist. See the
documentation of the Spam-X plugin for details.