Rationalize changes-entry and commit-changes-entry.
changes-entry now does cvs update (and cvs edit if needed), and adds
the line.  This target will behave the same way (with extra CVS
traffic) if the file is up to date, and will be omre useful if
CHANGES-YYYY is not up to date.

commit-changes-entry depends on changes-entry and also commits.  This
target has no behavior changes.

Rename _CCE_FOO variables to PKGSRC_CHANGES_FOO, and merge with
existing varables.

We may need a changes-entry-no-update, or a new name for what
changes-entry does now.  Given how infrequently CHANGES-YYYY is up to
date when I want to add, I believe the new behavior will be welcome to
almost everyone.