Update to 0.40.1:
0.40.1:
This is a corrective release to the 0.40.0 release. A bug in which
the library would not update changed files on a rescan was fixed.
MB metadata reading was made fully compliant with the latest
specifications. The dependency on libmp4v2 was dropped due to
problems, and appropriate code was put in place instead. The
preferences dialog was made GNOME HIG compliant.

0.40.0:
A rewrite of the library backend, reorganization of the UI, improved
Last.fm backend and tighter integration of Last.fm within the entire
application, enhanced and debugged Podcast backend with newly added
support for video podcasts using Gstreamer, initial Jamendo support,
and a universal bookmarking system. The stability of the 0.40 series
is incomparably higher than the 0.36 versions. The application's
usability was also improved to a great degree.