Initial import of atf-0.1:
The Automated Testing Framework (ATF) is a collection of libraries and
utilities designed to ease unattended application testing in the hands of
developers and end users of a specific piece of software.

As regards developers, ATF provides the necessary means to easily create
test suites composed of multiple test programs, which in turn are a
collection of test cases.  It also attempts to simplify the debugging of
problems when these test cases detect an error by providing as much
information as possible about the failure.

As regards users, it simplifies the process of running the test suites and,
in special, encourages end users to run them often: they do not need to
have source trees around nor any other development tools installed to be
able to certify that a given piece of software works on their machine as
advertised.

Yes, these are (part of) the results of my SoC 2007 project :-)