Imported jdom.
There is no compelling reason for a Java API to manipulate XML to be
complex, tricky, unintuitive, or a pain in the neck. JDOM[TM] is both
Java-centric and Java-optimized. It behaves like Java, it uses Java
collections, it is completely natural API for current Java developers,
and it provides a low-cost entry point for using XML.

While JDOM interoperates well with existing standards such as the Simple
API for XML (SAX) and the Document Object Model (DOM), it is not an
abstraction layer or enhancement to those APIs. Rather, it provides a
robust, light-weight means of reading and writing XML data without the
complex and memory-consumptive options that current API offerings
provide.