Time::Format version 1.02 =========================
Time::Format provides a very easy way to format dates and times.  The
formatting functions are tied to hash variables, so they can be used
inside strings as well as in ordinary expressions.  The formatting
codes used are meant to be easy to remember, use, and read.  They
follow a simple, consistent pattern.  If I've done my job right, once
you learn the codes, you should never have to refer to the
documentation again.  A quick-reference page is included, just in
case.  ;-)

Time::Format can also format DateTime objects, and strings created
with Date::Manip.

Also provided is a tied-hash interface to POSIX::strftime and
Date::Manip::UnixDate.