Add audio/herrie, an ncurses-based audio player.
Blurb:

Herrie is a command line music player. It supports multiple audio file
formats (including MP3 and Ogg Vorbis) and  has  the  ability  to  send
playback  statistics  to AudioScrobbler. The word `herrie' is Dutch for
`clamor'.

Herrie uses a split-screen user interface: a playlist at the top and  a
filebrowser  at the bottom. The application has two modes of operation,
called `party mode' and `XMMS mode'. When the application is started in
party  mode,  the first song in the playlist gets started. It will also
be removed from the playlist as well. When the application  is  started
in  XMMS mode, any song can be started and will not be removed from the
playlist.  The default mode is party mode.