Importing Plone 3.0.4 as plone3.
This package is based on new zope's framework.

Plone is a ready-to-run content management system that is built on the
powerful and free Zope application server. Plone is easy to set up,
extremely flexible, and provides you with a system for managing web
content that is ideal for project groups, communities, web sites,
extranets and intranets.

Plone 3 runs on Zope 2.10.x and has many improved features from Plone 2.5.

   1. Inline editing
   2. Working Copy support
   3. Link and reference integrity checking
   4. Automatic locking and unlocking
   5. Easy collaboration and sharing
   6. Versioning, history and reverting content
   7. Upgraded visual HTML editor
   8. Powerful workflow capabilities
   9. Flexible authentication back-end
  10. Full-text indexing of Word and PDF documents
  11. Collections
  12. Presentation mode for content
  13. Support for the search engine Sitemap protocol
  14. Support for multiple mark-up formats
  15. Wiki support
  16. Automatic previous/next navigation
  17. Rules engine for content
  18. Auto-generated tables of contents
  19. Portlets engine
  20. Professional support, development, hosting & training