Import urlgrabber-3.0.0 as www/urlgrabber.
urlgrabber is a pure python package that drastically simplifies
the fetching of files. It is designed to be used in programs that
need common (but not necessarily simple) url-fetching features.
It is extremely simple to drop into an existing program and provides
a clean interface to protocol-independant file-access. Best of all,
urlgrabber takes care of all those pesky file-fetching details,
and lets you focus on whatever it is that your program is written
to do!