Added www/p5-Catalyst-Controller-FormBuilder version 0.04
This base controller merges the functionality of CGI::FormBuilder
with Catalyst and the following templating systems: Template Toolkit,
Mason and HTML::Template. This gives you access to all of FormBuilder's
niceties, such as controllablefield stickiness, multilingual support,
and Javascript generation. For more details, see CGI::FormBuilder
or the website at:

    http://www.formbuilder.org