Update net/unison to 2.27.53 Enable full destdir support.
Changes since 2.17:
* Several small fixes to the GTK2 UI to make it work better under Windows
* The backup functionality has been completely rewritten. The external
  interface has not changed, but numerous bugs, irregular behaviors, and
  cross-platform inconsistencies have been corrected.
* Some important safety improvements:
  o Added a new mountpoint preference, which can be used to specify a path
    that must exist in both replicas at the end of update detection
  o The confirmation of “big deletes” is now controlled by a boolean
preference confirmbigdeletes. Default is true, which gives the same behavior
as previously. (This functionality is at least partly superceded by the
mountpoint preference, but it has been left in place in case it is useful to
some people.)
  o If Unison is asked to “follow” a symbolic link but there is nothing at
the other end of the link, it will now flag this path as an error, rather
than treating the symlink itself as missing or deleted. This avoids a
potentially dangerous situation where a followed symlink points to an
external filesystem that might be offline when Unison is run (whereupon
Unison would cheerfully delete the corresponding files in the other
replica!).
* Smaller changes:
  o Added forcepartial and preferpartial preferences, which behave like force
and prefer but can be specified on a per-path basis. [Thanks to Alan Schmitt
for this.]
  o A bare-bones self test feature was added, which runs unison through some
of its paces and checks that the results are as expected. The coverage of the
tests is still very limited, but the facility has already been very useful in
debugging the new backup functionality (especially in exposing some subtle
cross-platform issues).
  o Refined debugging code so that the verbosity of individual modules can be
controlled separately. Instead of just putting '-debug verbose' on the
command line, you can put '-debug update+', which causes all the extra
messages in the Update module, but not other modules, to be printed. Putting
'-debug verbose' causes all modules to print with maximum verbosity.
  o Removed mergebatch preference. (It never seemed very useful, and its
semantics were confusing.)
  o Rewrote some of the merging functionality, for better cooperation with
external Harmony instances.
  o Changed the temp file prefix from .# to .unison.
  o Compressed the output from the text user interface (particularly when run
with the -terse flag) to make it easier to interpret the results when Unison
is run several times in succession from a script.
  o Changed the order of arguments to the default diff command (so that the +
and - annotations in diff's output are reversed).
  o Small fix to ctime (non-)handling in update detection under windows with
fastcheck.