Import docker-1.5 as x11/docker.
Docker is a docking application (WindowMaker dock app) which acts as a system
tray for KDE and GNOME2. It can be used to replace the panel in either
environment, allowing you to have a system tray without running the KDE/GNOME
panel or environment.

Docker is designed to work with Openbox 2, but it should work fine in any
window manager that supports WindowMaker dock apps.

If you running WindowMaker or otherwise and experiencing problems, you should
try using the -wmaker option to make docker keep a fixed size.