Update to 1.2.8
Unfortunately details changes don't seem to be available:

After a long hiatus, new releases of libspf2! Development work from a number
of contributors has resulted in the release of versions up to 1.2.8. It is
recommended that all users upgrade as soon as possible, since this release
fixes some significant bugs. This release also comes with new doxygen
documentation which should make the code easier to understand and
cross-reference.

Includes the fix for CVE-2008-2469