Import TECkit.
TECkit is a low-level toolkit intended to be used by other
applications that need to perform encoding conversions (e.g., when
importing legacy data into a Unicode-based application).  The primary
component of the TECkit package is therefore a library that performs
conversions; this is the "TECkit engine".  The engine relies on
mapping tables in a specific binary format (for which documentation is
available); there is a compiler that creates such tables from a
human-readable mapping description (a simple text file).