Import ffmpegthumbnailer version 1.3.0.
This video thumbnailer can be used by file managers to create thumbnails for
your video files. The thumbnailer uses ffmpeg to decode frames from the video
files, so supported videoformats depend on the configuration flags of ffmpeg.

This thumbnailer was designed to be as fast and lightweight as possible. The
only dependencies are ffmpeg, libpng and libjpeg.