Add package option 'doc'. If and only if this option is set the documentation for the tidy library is built with doxygen and then installed.
This new package option is set by default hence preserving the
previous package "behaviour" in the default case.

Not setting this option provides smaller build dependencies: doxygen
is not required.

Please note that the manual page of the command line tool tidy is
still provided even if the package option 'doc' is not set.

Bump PKGREVISION to 1.