Import tex-rsfs-2008.11620 as fonts/tex-rsfs.
The fonts provide uppercase `formal' script letters for use as symbols
in scientific and mathematical typesetting (in contrast to the
informal script fonts such as that used for the `calligraphic' symbols
in the TeX maths symbol font).  The fonts are provided as MetaFont
source, and as derived Adobe Type 1 format.  LaTeX support, for using
these fonts in mathematics, is available via one of the packages
calrsfs and mathrsfs.