Linux doesn't define PPP_FCS() in <net/ppp_defs.h> for userland.  pppdump.c expects PPP_FCS() to be defined however, and compilation therefore breaks in the presence of the previous version of this patch.
This version of the patch reverts back to the behaviour of the distribution
code for non-NetBSD systems, but retains the Pkgsrc patched behaviour of
pulling in the system copy of the header file for NetBSD and DragonFly.

This should only affect success or failure of the package build, so no
PKGREVISION bump.