Importing devel/p5-Log-Report version 0.26 as dependency of scheduled update of devel/p5-Hash-Case to 1.01{0}.
Handling messages to users can be a hassle, certainly when the same module
is used for command-line and in a graphical interfaces, and has to cope
with internationalization at the same time; this set of modules tries to
simplify this. Log::Report combines gettext features with Log::Dispatch-like
features. However, this module can be used to do only translations or only
message dispatching.