Import tex-harvard-2.0.5
This is a re-implementation, for LaTeX 2e, of the original
Harvard package. The bundle contains the LaTeX package, several
BibTeX styles, and a Perl package for use with
LaTeX2HTML. Harvard is an author-year citation style (all but
the first author are suppressed in second and subsequent
citations of the same entry); the package defines several
variant styles: apsr.bst for the American Political Science
Review; agsm.bst for Australian Government publications;
dcu.bst from the Design Computing Unit of the University of
Sydney; kluwer.bstwhich aims at the format preferred in
Kluwer publications; nederlands.bst which deals with
sorting Dutch names with prefixes (such as van) according to
Dutch rules; together with several styles whose authors offer
no description of their behaviour.