Note initial addition of p5-Dancer version 1.175 as www/p5-Dancer into The NetBSD Packages Collection.
Dancer is a web application framework designed to be as effortless
as possible for the developer, taking care of the boring bits as
easily as possible, yet staying out of your way and letting you get
on with writing your code.

Dancer aims to provide the simplest way for writing web applications,
and offers the flexibility to scale between a very simple lightweight
web service consisting of a few lines of code in a single file, all
the way up to a more complex fully-fledged web application with
session support, templates for views and layouts, etc.