Added devel/p5-MooseX-FollowPBP version 0.02
This module does not provide any methods. Simply loading it changes
the default naming policy for the loading class so that accessors
are separated into get and set methods. The get methods are prefixed
with "get_" as the accessor, while set methods are prefixed with
"set_". This is the naming style recommended by Damian Conway in
Perl Best Practices.

If you define an attribute with a leading underscore, then both the
get and set method will also have an underscore prefix.

If you explicitly set a "reader" or "writer" name when creating an
attribute, then that attribute's naming scheme is left unchanged.