Update sysutils/fabric to 1.0.1
Upstream changes:

Bugfixes

    * #301: Fixed a bug in local?s behavior when capture=False and output.stdout
(or .stderr) was also False. Thanks to Chris Rose for the catch.
    * #310: Update edge case in put where using the mode kwarg alongside
use_sudo=True runs a hidden sudo command. The mode kwarg needs to be octal but
was being interpolated in the sudo call as a string/integer. Thanks to Adam
Ernst for the catch and suggested fix.
    * #311: append was supposed to have its partial kwarg's default flipped from
True to False. However, only the documentation was altered. This has been fixed.
Thanks to Adam Ernst for bringing it to our attention.
    * #312: Tweak internal I/O related loops to prevent high CPU usage and poor
screen-printing behavior on some systems. Thanks to Kirill Pinchuk for the
initial patch.
    * #320: Some users reported problems with dropped input, particularly while
entering sudo passwords. This was fixed via the same change as for #312.

Documentation

    * Added a missing entry for env.path in the usage documentation.