Import py-zim-0.52 as editors/zim
Zim aims to bring the concept of a wiki to your desktop. Every page
is saved as a text file with wiki markup. Pages can contain links
to other pages, and are saved automatically. Creating a new page
is as easy as linking to a non-existing page. This tool is intended
to keep track of TODO lists or to serve as a personal scratch book.
But it will also serve you when writing longer and more complicated
documents.