Update ruby-dm-core pacakge to 1.1.0.
DataMapper 1.1.0 released

I¡Çm pleased to announce that we have released DataMapper 1.1.

This has been one of the most enjoyable releases in recent memory. The
community rallied together and compared to the previous release we had at
least 3-4x more people submitting patches and working together to get this
release ready.

DataMapper 1.1 brings several minor API changes, warranting the minor version
bump, and closes 52 tickets in Lighthouse. There have been many performance
improvements, some closing bottlenecks that result in as much as a 20x speedup
from the 1.0.2 behaviour.

As part of the bug fixing process we¡Çve refactored some of the objects we use
internally to group relationships and dependencies and removed methods and
classes that were deprecated in 1.0.

For more changes in detail, please refer:
http://datamapper.org/articles/datamapper-110-released.html.