Initial import of CPAN-Meta-2.112150.
Software distributions released to the CPAN include a META.json
or, for older distributions, META.yml, which describes the
distribution, its contents, and the requirements for building and
installing the distribution. The data structure stored in the
META.json file is described in CPAN::Meta::Spec.

CPAN::Meta provides a simple class to represent this distribution
metadata (or distmeta), along with some helpful methods for
interrogating that data.