Import Threading Building Blocks 3.0 Update 6 as parallel/threadingbuildingblocks.
Intel(r) Threading Building Blocks (Intel TBB) offers a rich and
complete approach to expressing parallelism in a C++ program.
It is a library that helps you take advantage of multi-core
processor performance without having to be a threading expert.
Intel TBB is not just a threads-replacement library. It represents
a higher-level, task-based parallelism that abstracts platform
details and threading mechanisms for scalability and performance.