Initial import of EB-Garamond-0.012:
Garamont's fonts have set a milestone, on which font designers have
been recurring ever since. EB Garamond is an open source project
to create a revival of Claude Garamont's famous humanist typeface
from the mid-16th century. Its design reproduces the original design
by Claude Garamont: The source for the letterforms is a scan of a
specimen known as the "Berner specimen", which, composed in 1592
by Conrad Berner, son-in-law of Christian Egenolff and his successor
at the Egenolff print office, shows Garamont's roman and Granjon
italic fonts at different sizes. Hence the name of this project:
Egenolff-Berner Garamond.

The Garamonds are probably the most copied typefaces in the world.
There are indeed lots of excellent Garamond fonts, also such that
try to approach the original in the same way as EB Garamond does.
In the world of free software, however, only few Garamond-inspired
fonts exist, and as far as I know, none with the scope of EB
Garamond. I know that competition is hard in this field, and these
fonts won't be able to stand up to their commercial counterparts
for quite some time. Nevertheless, it's time for the opensource
community to have a classical Garamond and I promise, I'll try hard
to learn and give my best.