Update tiff to 4.0:
At long last, libtiff 4.0.0 is finally released.  Libtiff 4.0.0 is the
successor to the libtiff 3.9.X release series.  It is intended to be
largely API compatible with the 3.9.X releases, but it is definitely
not ABI compatible so any software which plans to use it will need to
be recompiled.  With appropriate care, source code can easily compile
with both the 3.9.X releases and libtiff 4.0.0.

This release supports the BigTIFF TIFF format in which all offsets are
unsigned 64-bit, supporting huge files.  APIs which deal with tag
offsets are necessarily updated to pass 64-bit values.  I/O functions
supporting the TIFFClientOpen() interface are updated to pass 64-bit
offset values.