Update to GCC 4.4.7
GCC 4.4.7 is a bug-fix release containing fixes for regressions and serious
bugs in GCC 4.4.6.  This release marks the end of the maintainance of
the GCC 4.4 series.

This is the list of problem reports (PRs) from GCC's bug tracking system
that are known to be fixed in the 4.4.7 release. This list might not be
complete (that is, it is possible that some PRs that have been fixed are
not listed here).
http://gcc.gnu.org/bugzilla/buglist.cgi?bug_status=RESOLVED&resolution=FIXED&target_milestone=4.6.2