Import py27-flakes-0.5.0nb1 as devel/py-flakes from wip/py-flakes
Pyflakes is similar to PyChecker in scope, but differs in that it does
not execute the modules to check them. This is both safer and faster,
although it does not perform as many checks. Unlike PyLint, Pyflakes
checks only for logical errors in programs; it does not perform any
checks on style