Add three meta-packages, bulk-small, bulk-medium, and bulk-large. These are lists of packages that can be used to do restricted bulk builds on small/slow machines without having to spend time researching which packages to include.
They are supposed to take roughly 1 hour, 8 hours, and 24 hours on the
arbitrarily chosen yardstick of my build machine; my hope is that this
translates into only moderately crazy lengths of time for slow
platforms where a full bulk build takes 6-8 months.

The lists will doubtless need some adjustment to meet this time
budget. Also, please take any discussion of which packages are and are
not included to tech-pkg or pkgsrc-users. The intent is to provide a
maximally useful set of a given size, but that's naturally difficult
to define.