Adding package for CPAN distribution Lingua-EN-Words2Nums version 0.18 into textproc/p5-Lingua-EN-Words2Nums.
This module converts English text into numbers. It supports both ordinal
and cardinal numbers, negative numbers, and very large numbers.

The main subroutine, which is exported by default, is words2nums(). This
subroutine, when fed a string, will attempt to convert it into a number.
If it succeeds, the number will be returned. If it fails, it returns undef.