Bump 1.8 to 4.0.4, based on PR pkg/48297 (by  Niclas Rosenvik: nicro465 at student.liu.se, also with PR pkg/45521 to 3.1 by kamelderouiche at yahoo.com)
Release Note since 4.0 from:
  http://freecode.com/projects/qtexmaker/releases/357375
--------------------
4.0.4     30 Aug 2013 16:55
      Support for the 0.24 poppler version has been
added, and a compilation problem on ARM architecture has been
fixed. Dragging and dropping image files in the editor is now
supported.

4.0.3     30 Jul 2013 12:46
      The Windows and Mac OS X Lion versions are now
compiled with Qt 5.1 and poppler 0.22. Find-and-replace
actions can now be applied to a selection only. A tool that
converts source files to Unicode has been added. Users can now
check if a new version is available via a new option in the
"Help" menu (less)

4.0.2     09 May 2013 11:33
      Mac OS X high DPI support has been added to the
PDF viewer, the symbols, and almost all of the icons. The
part/chapter/section tags are now bold in the editor. On
Linux, the session file is now stored in the standard ./cache
directory instead of the /tmp folder. New translations have
been added (less)

4.0.1     16 Mar 2013 13:47
      A potential crash when synctex sent a bad page
number has been fixed. The -Pdownload35 option has been added
to the default dvips command for MiKTeX users on Windows.

4.0     08 Mar 2013 22:07
      The code has been ported to Qt5 (but this
version can also be compiled with Qt4). From this version,
Texmaker will be principally developed under Qt5. A keyboard
shortcut has been added to toggle between the master and the
current document. A specific "quick build" command has been
padded for the .asy files: the asymptote figures can now be
compiled and displayed with the F1 shortcut.