Import py27-acora-1.8 as textproc/py-acora.
Acora is 'fgrep' for Python, a fast multi-keyword text search engine.

Based on a set of keywords, it generates a search automaton (DFA) and runs it
over string input, either unicode or bytes.

It is based on the Aho-Corasick algorithm and an NFA-to-DFA powerset
construction.

Acora comes with both a pure Python implementation and a fast binary module
written in Cython. However, note that the current construction algorithm is not
suitable for really large sets of keywords (i.e. more than a couple of
thousand).