Import poppler-qt5-0.26.0 as print/poppler-qt5, based on wip package by nros.
This is the QT5 wrapper for poppler, a PDF rendering library. It
is a fork of the xpdf 3.0 PDF viewer developed by Derek Noonburg
of Glyph and Cog, LLC.

Poppler provides PDF rendering functionality as a shared library,
to centralize the maintenence effort and to reduce duplicated
effort. And the poppler developers would like to move libpoppler
forward in a number of areas that doesn't fit within the goals of
xpdf.