Update to 1.5.7
Upstream changes:
1.5.7 (2014-05-29)

Bug fixes:
Skip the traceback on "Broken pipe" signal. (Issue #275)
Do not exit when an option in setup.cfg or tox.ini is not recognized.
Check the last line even if it does not end with a newline. (Issue #286)
Always open files in universal newlines mode in Python 2. (Issue #288)