Import webkit-gtk3-1.10.2nb12 as www/webkit1-gtk3 to match webkit1-gtk.
WebKit is an open source web browser engine. WebKit is also the name of
the Mac OS X system framework version of the engine that's used by
Safari, Dashboard, Mail, and many other OS X applications. WebKit's HTML
and JavaScript code began as a branch of the KHTML and KJS libraries
from KDE.

This is the GTK3+ port of major version 1 of the engine.