Make pgsql option defult to off.
This option fixes a pgsql version, and we're not really set up to deal
with that.  It makes postgis only build for that one version.  So far,
I'm not aware of a reason to have pgsql support in gdal.  So, just
turn it off, and we can figure out a way to deal if there is actually
a need.

Discussed with Filip during postgis2 packaging, and I think suggested
to Brook and Adam.