Initial import of highr v0.3 as textproc/R-highr.
This package provides syntax highlighting for R source code. Currently
it supports LaTeX and HTML output. Source code of other languages can
be supported via Andre Simon's Highlight package.

Note: this commit is part of reorganizing some of the recently
imported R packages, which are being reimported into more appropriate
categories (and removed from math) as a result of a recent discussion
on tech-pkg and privately with wiz@.  See the thread starting with:

  http://mail-index.netbsd.org/tech-pkg/2014/09/05/msg013558.html