Update ruby193-base and related packages to 1.9.3p550 (Ruby 1.9.3-p550).
Ruby 1.9.3-p550 Released

We are pleased to announce the release of Ruby 1.9.3-p550.

This release includes a security fix for DoS vulnerability of REXML.

    * CVE-2014-8080: Denial Of Service XML Expansion

This release also includes the change of default settings of
ext/openssl. Insecure SSL/TLS options are now turn off by default.

    * Changed default settings of ext/openssl

And, in addition, bandled jQuery for darkfish template of RDoc is also
updated.