Import ioquake3-raspberrypi-20150118 as games/ioquake3-raspberrypi.
ioquake3 is a game engine project which aims to build upon id Software's Quake
3 source code release. The source code was released on August 20, 2005 under
the GPL.

The project's goal is to create a bug-free, enhanced open source Quake 3
source code distribution upon which new games and projects can be based, as
well as an improved version (in relation to the original binaries) of the
engine for playing Quake III: Arena, the Team Arena expansion pack and all the
popular mods.

This version of the package is optimized for the Raspberry Pi.