Import retroarch-0.0.20150203 as emulators/retroarch.
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

RetroArch is the official reference frontend for the libretro API.