Import libretro-stella-20150205 as emulators/libretro-stella.
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

Stella is an Atari 2600 video game system emulator.