Import p5-Signal-Mask-0.008 as devel/p5-Signal-Mask.
Signal::Mask is an abstraction around your process or thread signal
mask. It is used to fetch and/or change the signal mask of the calling
process or thread.
The signal mask is the set of signals whose delivery is currently blocked
for the caller. It is available as the global hash %Signal::Mask.