lang/gcc-aux: Fix build on NetBSD 7.0-Beta
Like lang/gcc5-aux, gcc-aux was failing to build on NetBSD 7.0-beta
due to a "operation not permitted" error emitted by pkgsrc gold linker.

The root cause isn't known (problem with NetBSD 7 full-stop?  problem
with binutils, but only NetBSD7?).  However, the base binutils which
is version 2.23 is sufficient (barely) for gcc 4.9.

The build failure is solved by using base binutils for NetBSD 7*.
No revbump required because there is no change for platforms on which
gcc-aux has already potentially been built.

NOTE: As noted on gcc5-aux, stack-checks that passed on NetBSD 5 and 6
are now failing on GNAT on NetBSD 7.  This is because the signal
trampoline has apparently changed and the pattern the unwinder was
looking for is never found, resulting in a segfault instead of a frame
unwind.  The fix would be to determine the new signal trampoline
pattern and have it switch based on param.h version (or something like
that).