Update to CHICKEN 4.10.0
This is mostly a bugfix release, containing quite a lot of accumulated
bug fixes but relatively few new features.  Nevertheless, there are a
great many changes in this release since 4.9.0.  It has three security
bugfixes: for CVE-2014-6310, CVE-2014-9651 and CVE-2015-4556.

Another important security-related cleanup/improvement is related to path
expansion: many file procedures in CHICKEN used to automatically and
implicitly convert paths containing tilde ("~") characters to $HOME,
and expand shell variables.  This behaviour was deprecated in 4.9.0,
and has now been removed, in favor of the "pathname-expand" egg.

A few of the build variables have been renamed, so if you have a custom
build script, please review the NEWS file carefully.

Other notable changes:

* The performance of reading and writing SRFI-4 u8vectors has been
   substantially improved.
* Various bugs in the functors implementation have been fixed, making
   them better usable.
* chicken.h can once more be correctly compiled with a C++ compiler.
* The -r5rs-syntax flag has been fixed to enforce r5rs syntax.

For the full list of changes, see bundled NEWS or
http://code.call-cc.org/releases/4.10.0/NEWS-4.10.0