Bring xemacs-packages into the 21st century.
o Update packages

o Use DESTDIR features for installing, getting rid of the need to be
root and gaining the ability to patch elisp files in the process.

o Add a mechanism to re-byte-compile the patched elisp files, and
patches to deal with non-utf-8 XEmacsen.

o Grab maintainership

Tested with both 21.4 and 21.5 packages.