Re-Import keepassx-0.4.4 as security/keepassx0 before update to new major version.
KeePass is a free/open-source password manager or safe which helps you to
manage your passwords in a secure way. You can put all your passwords in one
database, which is locked with one master key or a key-disk. So you only have
to remember one single master password or insert the key-disk to unlock the
whole database. The databases are encrypted using the best and most secure
encryption algorithms currently known (AES and Twofish).

This package contains v0.4.x of the program with KeePass v1 database
format (kdb) support.