Import freecol-0.11.6 as games/freecol.
FreeCol is a turn-based strategy game based on the old game
Colonization, and similar to Civilization. The objective of the
game is to create an independent nation.  You start with only a
few colonists defying the stormy seas in their search for new land.
Will you guide them on the Colonization of a New World?