Import i3status-2.10 as sysutils/i3status
i3status is a small program (about 3000 SLOC) for generating a status bar
for dzen2, xmobar or similar programs. It is designed to be very efficient
by issuing a very small number of system calls, as one generally wants
to update such a status line every second.
This ensures that even under high load, your status bar is updated correctly.
Also, it saves a bit of energy by not hogging your CPU as much as spawning
the corresponding amount of shell commands would.

Originally packaged in pkgsrc-wip by:
- Pierre Allegraud,
- Vitaly Shevtsov,
- Niclas Rosenvik,
- Mateusz Poszwa,
- myself,
with maintanance commits by:
- Aleksej Saushev.