New package, go-termbox.
Termbox is a library that provides a minimalistic API which allows the
programmer to write text-based user interfaces. The library is crossplatform
and has both terminal-based implementations on *nix operating systems and a
winapi console based implementation for windows operating systems. The basic
idea is an abstraction of the greatest common subset of features available on
all major terminals and other terminal-like APIs in a minimalistic fashion.
Small API means it is easy to implement, test, maintain and learn it, that's
what makes the termbox a distinct library in its area.