Merge my recent fixes to emacs20 and emacs21:    - set ELFSIZE correctly (I still have no idea how this ever worked)    - insert workaround for gcc5's calloc bug
For some reason I thought emacs22 wasn't affected by these issues, or
I would have done this earlier.

This does not make the package build on -current, but it gets a lot
farther now - it segvs building the quail elisp. I have a feeling
we've seen this problem before also, but if so it was some years back
and I can't remember what it was about.