Import p5-Eval-LineNumbers-0.1 as devel/p5-Eval-LineNumbers.
Add a "#line "this-file" 392" comment to heredoc/hereis text that is
going to be eval'ed so that error messages will point back to the right
place.

Please note: when you embed "\n" in your code, it gets expanded in
double-quote hereis documents so it will mess up your line numbering.
Use "\\n" instead when you can.