Import acme-tiny's latest commit (versioned as 0.0.20170214)
This is a tiny, auditable script that you can throw on your server to issue and
renew Let's Encrypt certificates. Since it has to be run on your server and
have access to your private Let's Encrypt account key, I tried to make it as
tiny as possible (currently less than 200 lines). The only prerequisites are
python and openssl.

Tested successfully on NetBSD 7.0.2 (i386).

"please import!" wiz@