Add ruby-gettext-setup package version 0.25.
gettext-setup gem

This is a simple gem to set up i18n for Ruby projects (including Sinatra web
apps) using gettext and fast gettext.

This project sets the default locale to English. If the user has set a
different locale in their browser preferences, and we support the user's
preferred locale, strings and data formatting will be customized for that
locale.