Update to 2.8.6
Upstream changes:
2017 July 21

GNU nano 2.8.6 "Kekulé" offers a new feature: the ability
to do softwrapping between words -- at whitespace --
instead of always at the edge of the screen.  This can
be activated with -a or --atblanks or 'set atblanks'
together with the softwrap option.  This release further
fixes a handful of rare display glitches, fixes a build
failure on AIX, harmonizes the quoting rules in the rc
files, and renames the option 'cut' to 'cutfromcursor'
(please update your nanorc files before 2020).