OpenJPEG 2.2.0:
This new release includes a significant number of improvements and bug fixes. In particular:
* Multi-threading support at decoding side
* Several speed optimisations both at encoder and decoder, and both on Wavelet Transform and Entropy Coding parts. On our test set, a single-threaded execution is now around 20% faster (encoding or decoding).
* Huge memory consumption reduction at decoding side (~60% reduction on large images)
* Several important bug fixes, in particular the one that was preventing OpenJPEG to encode lossless in some specific situations, as well as those related to mode switches (BYPASS/LAZY, RESTART/TERMALL, etc).
* Several security fixes thanks to the inclusion of OpenJPEG in the Google OSS Fuzz project.

Beside that, several improvements have been brought to the project maintenance, like inclusion of benchmarking scripts to compare speed with latest available kakadu binaries.

And, last but not least, API and ABI are 100% compatible with previous version.