Import tmate version 2.2.1 as misc/tmate.
tmate provides an instant pairing solution, allowing you to share a
terminal with one or several teammates.  Together with a voice call, it's
almost like pairing in person.  The terminal sharing works by using SSH
connections to backend servers maintained by tmate upstream developers.
Teammates need to be given a randomly-generated token to be able to join a
session.

tmate is a modified version of tmux, and uses the same configurations such
as keybindings, color schemes etc.