Update slibtool to 0.5.17
0.5.17:

This patch release fixes a subtle bug in the newly added support for
packages that depend on the reordering of linker arguments.

0.5.16:

This patch release fixes a subtle bug in slibtool-shared's
generation of dependency lists when creating convenience
library archives. Beginning with this release, the mesa
package can be built with slibtool in its entirety.

- slibtool-shared: properly handle .la arguments when creating an archive.
- slbt_exec_link_create_dep_file(): properly initialize and track fdeps.

0.5.15:

This patch release provides seamless support for the building of
packages which rely on the ordering of linker arguments. With
slibtool, reordering takes place only when deemed necessary.

- link mode: support packages that rely on argument reordering.
- error trace: remove and fix inadvertent calls to strerror(3).

0.5.14:

This patch release provides several fixes at the driver level, as well as
several improvements to slibtool's own internals. Addition of extended
variable support to slibtool's ./configure now makes it fully compatible
with (distro) build systems that pass those variable to ./configure by
default.

- link mode: fix fdepfile construction when determining dynlib dependency.
- link mode: static archives: always have ar(1) act as ranlib(1).
- driver: refine handling of --target.
- build system: added extended dir variables.