tex-makecell{,-doc}: import version 0.1e
This package supports common layouts for tabular column heads
in whole documents, based on one-column tabular environment. In
addition, it can create multi-lined tabular cells. The Package
also offers: a macro which changes the vertical space around
all the cells in a tabular environment (similar to the function
of the tabls package, but using the facilities of the array)
macros for multirow cells, which use the facilities of the
multirow package; macros to number rows in tables, or to skip
cells; diagonally divided cells; horizontal lines in tabular
environments with defined thickness.