libpciaccess: update to 0.14.
This release adds support for 32-bit PCI domain IDs, which are
apparently common in virtualized environments. This support has been
tested on Linux, and probably also "just works" on FreeBSD and Solaris,
to the extent that the underlying OS supports it.

If you were allocating a struct pci_device yourself, or embedding it in
another structure, this release would be an ABI break. That said,
anyone doing this would already be violating the implicit API contract
(not to mention good taste) and we were unable to find any consumers of
this library misbehaving in this way. As a result the library's soname
does not change in this release.