yarn: Import version 1.3.2 into pkgsrc.
Yarn is a package manager for your code.  It allows you to use and share
code with other developers from around the world.  Yarn does this quickly,
securely, and reliably so you don't ever have to worry.

Yarn allows you to use other developers' solutions to different problems,
making it easier for you to develop your software.  If you have problems,
you can report issues or contribute back, and when the problem is fixed,
you can use Yarn to keep it all up to date.

Code is shared through something called a package (sometimes referred to
as a module).  A package contains all the code being shared as well as a
package.json file which describes the package.