Switch lang/php72 to use pkgsrc pcre library instead of the bundled one
PHP bundles it's own copy of pcre 8.41 with the distribution, which
is used when build doesn't find system one. Switch to the pkgsrc
pcre, so that we don't need to maintain local PHP JIT patch here.
Upcoming PHP 7.3 already switched to pcre2, so no value maintaining
private PCRE patch any more (if there ever was).

XXX devel/pcre doesn't seem to enable PCRE JIT, might be good to do so