graphics/R-viridisLite: Import version 0.3.0
Implementation of the 'viridis', 'magma', 'plasma', 'inferno', and
'cividis' color maps for 'R'. 'viridis', 'magma', 'plasma', and
'inferno' are ported from 'matplotlib', a popular plotting library for
Python. These color maps are designed in such a way that they will
analytically be perfectly perceptually-uniform, both in regular form
and also when converted to black-and-white. They are also designed to
be perceived by readers with the most common form of color blindness
(all color maps in this package) and color vision deficiency
('cividis' only). This is the 'lite' version of the more complete
'viridis' package.