Fix build with Go 1.10, bump revision.
It seems that the rules for pattern matching have changed at some point,
which is what is breaking these builds (or more exactly, their PLISTs).
Now, foo/... does not match anything under foo/vendor AFAICT, and packages
that do not match the pattern do not get .a files explicitly created.