print/py-weasyprint0: Import WeasyPrint 0.42.3.
NOTE: This package caters to Python 2.7 environments, support for
      which was dropped in subsequent releases.


WeasyPrint is a visual rendering engine for HTML and CSS that can export
to PDF. It aims to support web standards for printing.
WeasyPrint is free software made available under a BSD license.

It is based on various libraries but *not* on a full rendering engine like
WebKit or Gecko. The CSS layout engine is written in Python, designed for
pagination, and meant to be easy to hack on.