devel/bison: add patches to fix building with glibc >= 2.27
Relates to PR pkg/53826, further details are provided there. This is a
temporary fix: the newest release of Bison addresses this by including
a newer version of gnulib. I'm applying this less obtrusive change for
now. No PKGREVISION, because there should be no change to existing
packages, it addresses build failures only.