py-hypothesis: updated to 4.5.11
4.5.11:
This release fixes :issue:1813, a bug introduced in :ref:3.59.1 <v3.59.1>, which caused :py:meth:~hypothesis.strategies.random_module to no longer affect the body of the test: Although Hypothesis would claim to be seeding the random module in fact tests would always run with a seed of zero.

4.5.10:
This patch fixes an off-by-one error in the maximum length of :func:~hypothesis.strategies.emails. Thanks to Krzysztof Jurewicz for :pull:1812.

4.5.9:
This patch removes some unused code from the shrinker. There is no user-visible change.

4.5.8:
This release fixes an internal IndexError in Hypothesis that could sometimes be triggered during shrinking.

4.5.7:
This release modifies the shrinker to interleave different types of reduction operations, e.g. switching between deleting data and lowering scalar values rather than trying entirely deletions then entirely lowering.

This may slow things down somewhat in the typical case, but has the major advantage that many previously difficult to shrink examples should become much faster, because the shrinker will no longer tend to stall when trying some ineffective changes to the shrink target but will instead interleave it with other more effective operations.

4.5.6:
This release makes a number of internal changes to the implementation of :func:hypothesis.extra.lark.from_lark. These are primarily intended as a refactoring, but you may see some minor improvements to performance when generating large strings, and possibly to shrink quality.

4.5.5:
This patch prints an explanatory note when :issue:1798 is triggered, because the error message from Numpy is too terse to locate the problem.

4.5.4:
In Python 2, long integers are not allowed in the shape argument to :func:~hypothesis.extra.numpy.arrays. Thanks to Ryan Turner for fixing this.

4.5.3:
This release makes a small internal refactoring to clarify how Hypothesis instructs tests to stop running when appropriate. There is no user-visible change.

4.5.2:
This release standardises all of the shrinker's internal operations on running in a random order.

The main effect you will see from this that it should now be much less common for the shrinker to stall for a long time before making further progress. In some cases this will correspond to shrinking more slowly, but on average it should result in faster shrinking.

4.5.1:
This patch updates some docstrings, but has no runtime changes.

4.5.0:
This release adds exclude_min and exclude_max arguments to :func:~hypothesis.strategies.floats, so that you can easily generate values from open or half-open intervals