gettext-lib: Always add -lintl on SunOS.
This effectively turns on BROKEN_GETTEXT_DETECTION=yes on SunOS and adds
-lintl whenever a package pulls in gettext-lib.  The SunOS linker is
more strict than others and will ensure that library dependencies are
explicit.  Recent changes to gettext-lib and glib made the dependency
implicit and caused a lot of fallout.

Use BUILDLINK_LDFLAGS instead of BUILDLINK_LIBS so that it actually gets
added through buildlink, not only when packages honour ${LIBS}.