nghttp2: updated to 1.39.1
v1.39.1:

nghttpx
This release fixes the bug that log-level is not set with cmd-line or configuration file. It also fixes FPE with default backend.

v1.39.0:

lib
libnghttp2 now ignores content-length in 200 response to CONNECT request as per RFC 7230.

third-party
mruby has been upgraded to 2.0.1.

asio
libnghttp2-asio now supports boost-1.70.

src
http-parser has been replaced with llhttp.

nghttpx
nghttpx now ignores Content-Length and Transfer-Encoding in 1xx or 200 to CONNECT.

This release fixes the bug that the log level does not change to the default value on configuration reload if log-level option is missing in new configuration.