py-scipy: updated to 1.3.0
SciPy 1.3.0 Release Notes

SciPy 1.3.0 is the culmination of 5 months of hard work. It contains
many new features, numerous bug-fixes, improved test coverage and better
documentation. There have been some API changes
in this release, which are documented below. All users are encouraged to
upgrade to this release, as there are a large number of bug-fixes and
optimizations. Before upgrading, we recommend that users check that
their own code does not use deprecated SciPy functionality (to do so,
run your code with python -Wd and check for DeprecationWarning s).
Our development attention will now shift to bug-fix releases on the
1.3.x branch, and on adding new features on the master branch.

This release requires Python 3.5+ and NumPy 1.13.3 or greater.

For running on PyPy, PyPy3 6.0+ and NumPy 1.15.0 are required.

Highlights of this release
- Three new stats functions, a rewrite of pearsonr, and an exact
  computation of the Kolmogorov-Smirnov two-sample test
- A new Cython API for bounded scalar-function root-finders in scipy.optimize
- Substantial CSR and CSC sparse matrix indexing performance
  improvements
- Added support for interpolation of rotations with continuous angular
  rate and acceleration in RotationSpline


SciPy 1.2.0 Release Notes

SciPy 1.2.0 is the culmination of 6 months of hard work. It contains
many new features, numerous bug-fixes, improved test coverage and better
documentation. There have been a number of deprecations and API changes
in this release, which are documented below. All users are encouraged to
upgrade to this release, as there are a large number of bug-fixes and
optimizations. Before upgrading, we recommend that users check that
their own code does not use deprecated SciPy functionality (to do so,
run your code with python -Wd and check for DeprecationWarning s).
Our development attention will now shift to bug-fix releases on the
1.2.x branch, and on adding new features on the master branch.

This release requires Python 2.7 or 3.4+ and NumPy 1.8.2 or greater.

This will be the last SciPy release to support Python 2.7.
Consequently, the 1.2.x series will be a long term support (LTS)
release; we will backport bug fixes until 1 Jan 2020.

For running on PyPy, PyPy3 6.0+ and NumPy 1.15.0 are required.

Highlights of this release
- 1-D root finding improvements with a new solver, toms748, and a new
  unified interface, root_scalar
- New dual_annealing optimization method that combines stochastic and
  local deterministic searching
- A new optimization algorithm, shgo (simplicial homology
  global optimization) for derivative free optimization problems
- A new category of quaternion-based transformations are available in
  scipy.spatial.transform