py-pywavelets: updated to 1.0.3
PyWavelets 1.0.3 is functionally equivalent to the 1.0.2 release. It was made
to add the add an archive of the JOSS paper to the 1.0.x branch and serve as a
reference corresponding to the version of the software reviewed that was peer
reviewed.

PyWavelets 1.0.2 is a bug-fix and maintenance release with no new features
compared to 1.0.1.

PyWavelets 1.0.1 is a bug-fix release with no new features compared to 1.0.0.

We are very pleased to announce the release of PyWavelets 1.0.  We view this
version number as a milestone in the project's now more than a decade long
history.  It reflects that PyWavelets has stabilized over the past few years,
and is now a mature package which a lot of other important packages depend on.
A listing of those package won't be complete, but some we are aware of are:

- scikit-image <https://scikit-image.org>_ - image processing in Python
- imagehash <https://github.com/JohannesBuchner/imagehash>_ - perceptual image hashing
- pyradiomics <https://github.com/Radiomics/pyradiomics>_ - extraction of Radiomics features from 2D and 3D images and binary masks
- tomopy <https://github.com/tomopy/tomopy>_ - Tomographic Reconstruction in Python
- SpikeSort <https://github.com/btel/SpikeSort>_ - Spike sorting library implemented in Python/NumPy/PyTables
- ODL <https://github.com/odlgroup/odl>_ - operator discretization library

This release requires Python 2.7 or >=3.5 and NumPy 1.9.1 or greater.
The 1.0 release will be the last release supporting Python 2.7.  It will be a
Long Term Support (LTS) release, meaning that we will backport critical bug
fixes to 1.0.x for as long as Python itself does so (i.e. until 1 Jan 2020).