libesmtp: update to 1.0.6nb4.
Disable ntlm plugin. It is not built by default on NetBSD 9.0.
I read the configure output and script and couldn't determine
the problem. This makes the build succeed again, and someone
who is more interested in ntlm support can fix it.

For perspective, last update in 2010 and no upstream available.

Some pkglint cleanup while here.