mk: Ensure only the first NetBSD rcsid is used.
There are a couple of places in pkgsrc where rcsid's are generated, and
these were leaking into "pkg_info -b" output.  This had the side effect of
causing those packages and all their dependencies to be rebuilt for every
bulk build as the output and the respective files always mis-matched.