math/py-netCDF4: Import version 1.5.2
netCDF version 4 has many features not found in earlier versions of
the library, such as hierarchical groups, zlib compression, multiple
unlimited dimensions, and new data types. It is implemented on top of
HDF5. This module implements most of the new features, and can read
and write netCDF files compatible with older versions of the
library. The API is modelled after Scientific.IO.NetCDF, and should be
familiar to users of that module.

Packaged by Kamel Ibn Aziz Derouiche and updated by wiz@ and me.