ham/gnuradio-qt: Explain status relative to upstream and qt
The quick summary is that the current package is from April, and
upstream had a new release in August.  That new release switches from
qt4 to qt5.  Updating GNU Radio in pkgsrc is complex, but tends to
happen fairly soon after the upstream release.  As part of that
update, this package is expected to switch to qt5.