devel/p5-Test-CPAN-Meta-JSON: import p5-Test-CPAN-Meta-JSON-0.16
This module was written to ensure that a META.json file, provided with a
standard distribution uploaded to CPAN, meets the specifications that are
slowly being introduced to module uploads, via the use of package makers
and installers such as ExtUtils::MakeMaker, Module::Build and
Module::Install.
See CPAN::Meta for further details of the CPAN Meta Specification.