sysutils/bup: Workarounds for dealing with haskell, github, github.mk
github.mk presumes that there is one distfile, but bup has two because
it uses pandoc which uses haskell.  Set variables that should result
in fetching both of them.  Because this does not work, I have put a
copy of the man tarball at MASTER_SITE_BACKUP.

(No PKGREVISION++ because if you had the distfiles before you'll get
the same binary package now.)