lang/ruby27-base: add ruby-base version 2.7.0
Add ruby27-base 2.7.0 package.  This package is Ruby 2.7 release
minimum base package.


Ruby is the interpreted scripting language for quick and
easy object-oriented programming.  It has many features to
process text files and to do system management tasks (as in
Perl).  It is simple, straight-forward, and extensible.

Features of Ruby are shown below.

  + Simple Syntax
  + *Normal* Object-Oriented features (ex. class, method calls)
  + *Advanced* Object-Oriented features (ex. Mix-in, Singleton-method)
  + Operator Overloading
  + Exception Handling
  + Iterators and Closures
  + Garbage Collection
  + Dynamic Loading of Object files (on some architecture)
  + Highly Portable (works on many UNIX machines, and on DOS, Windows,
    Mac, etc.)

Ruby 2.7 introduces a number of new features and performance
improvements, most notably:

  * Pattern Matching (Experimental)
  * REPL improvement
  * Compaction GC
  * Separation of positional and keyword arguments