py-graphviz: updated to 0.14
Version 0.14
------------

Improve handling of escaped quotes (``\"``). Different from other layout engine
escapes sequences such as ``\l`` and ``\N`` (which are passed on as is by
default), there is no use case for backslash-escaping a literal quote character
because escaping of quotes is done by this library. Therefore, a
backslash-escaped quote (e.g. in ``label='\\"'``) is now treated the same as a
plain unescaped quote, i.e. both ``label='"'`` and ``label='\\"'`` produce
the same DOT source ``[label="\""]`` (a label that renders as a literal quote).
Before this change, use of ``'\\"'`` could break the quoting mechanism creating
invalid or unintended DOT, possibly leading to syntax errors from the rendering
process.

Add notebook section to documentation.

Add ``sphinx.ext.viewcode`` to docs (note that this currently lacks links for
methods, so that not all of the code is linked; use the source repo for reading
on).

Improve test and doc building config.