tnftp: Move editline support to a non-default option.
This code was very complicated and easily broke, for example when termcap is
builtin but curses is not, resulting in circular dependencies.

Such core packages should default to a safe set of build options, unless they
can be reliably added to PKG_SUGGESTED_OPTIONS, to avoid a miserable first
experience for users running on older or quirky platforms.