textproc/dwdiff: import dwdiff-2.1.2
dwdiff is a diff program that operates at the word level instead of the line
level. It is different from wdiff in that it allows the user to specify what
should be considered whitespace, and in that it takes an optional list of
characters that should be considered delimiters. Delimiters are single
characters that are treated as if they are words, even when there is no
whitespace separating them from preceding words or delimiters. dwdiff is mostly
commandline compatible with wdiff. Only the --autopager, --terminal and
--avoid-wraps options are not supported.

The default output from dwdiff is the new text, with the deleted and inserted
parts annotated with markers. Command line options are available to change
both what is printed, and the markers.

dwdiff is licensed under the GNU General Public License version 3. See the file
COPYING for details.