sysutils/zabbix*: Update to 4.0.21
As provided on pkgsrc-users by Yuuki Enomoto.

Packaging changes: Update MASTER_SITE.

Upstream lacks NEWS and the page that their NEWS file points to does
not contain NEWS.

A quick glance over upstream, and the fact that this is a series of
micro updates, leads one to guess that this update basically contains
minor improvements and bugfixes.