xf86-video-vesa: update to 2.5.0.
This release rolls up a few minor bug fixes since 2.4.0. We now refuse
to run on machines with UEFI firmware (on Linux only, patches welcome
for other OSes) since it won't work in the general case and you
probably have a kernel framebuffer driver running already. We also only
attempt to use 24bpp if the alternative would be pseudocolor, but note
that since xserver 1.20 24bpp will not work at all. If you'd like to
fix that case, please note that it is filed as issue #4:

https://gitlab.freedesktop.org/xorg/driver/xf86-video-vesa/-/issues/4

We also notice several cases of running on emulated GPUs, in which case
the shadow framebuffer is disabled by default to improve performance by
reducing the number of copies.

All users are recommended to upgrade, ideally to a better video card
and/or driver.

Full changelog:

Adam Jackson (5):
      Newline-terminate a few error messages
      Refuse to run on UEFI machines
      Try harder to avoid 24bpp
      Disable shadow by default on known-virtual GPUs
      vesa 2.5.0

Alan Coopersmith (2):
      Update README for gitlab migration
      Update configure.ac bug URL for gitlab migration

Benjamin Tissoires (1):
      Refuse to run on UEFI machines running older kernels