py-deepstate: import version 0.1
DeepState is a framework that provides C and C++ developers with a common
interface to various symbolic execution and fuzzing engines. Users can write
one test harness using a Google Test-like API, then execute it using multiple
backends without having to learn the complexities of the underlying engines. It
supports writing unit tests and API sequence tests, as well as automatic test
generation.

This package installs the executors for DeepState.