lang/rust: Adjust openssl bootstrap depends (NetBSD only)
On NetBSD, further limit the DEPENDS on openssl to i386.  It turns out
that the sparc64 and powerpc bootstraps are not linked against openssl
as I previously assumed.

No change to limiting this DEPENDS to 9+.  No change to any other OS.

This change is believed to fix rust building on NetBSD 9 sparc64.

As discussed on tech-pkg@.