audio/librespot: import
Open source client library for Spotify. It enables applications to use Spotify's
service to contol and play music via various backends, and to act as a Spotify
Connect receiver.
It is an alternative to the official and deprecated closed-source libspotify.
Additionally, it will provide extra features which are not available in the
official library.

Note: librespot only works with Spotify Premium. This will remain the case for
the foreseeable future, as we are unlikely to work on implementing the features
such as limited skips and adverts that would be required to make librespot
compliant with free accounts.