gettext-lib: fix built-in gettext detection on musl
Some packages, for example GNU sed, still use old versions of
gettext.m4 (serial 67 or older) despite having recent releases.

These versions of gettext.m4 fail to detect gettext on musl, which
has gettext built-in to libc. While there is some logic in
gettext-lib/builtin.mk for this situation, it only covers the case
where libintl is separate from libc. To fix this, set the corresponding
configure variables when gettext is built-in to libc.