security/doas: update to 6.3p2
6.3p2

This release introduces a new utility called vidoas (vi doas). This tool is a
shell script which creates a copy of the doas.conf file, allows the admin to
edit the file, and then checks its syntax for errors. If a problem is found,
vidoas reports which line the error was on and asks us to try editing the file
again. Once the new doas.conf file contains the proper syntax, it is installed
and overwrites the old doas.conf file.

This tool is designed to assist admins and avoid introducing errors to doas.conf
which might accidentally revoke admin access to the machine.

6.3p1
In this release, we work around a quirk of the GNU parameter parser which
required us to use double-dashes (--) after doas's parameters and before a
target command's parameters. In the past we used "doas -- pacman -Syu" and now
we can use simply "doas pacman -Syu".

This change affects only GNU/Linux systems, other platforms like FreeBSD,
NetBSD, etc already had this behaviour.

6.3
This release introduces a few minor changes:
-Added command line parameter (-S) which launches an interactive shell. This is
equivalent to "su -l" or "sudo -i".
-Updated documentation to include the new -S flag.
-Updated documentation to assist users in installing doas on some Linux
distributions, such as CentOS, that prevent PAM authentication from working by
default.

6.2p5
This release simply adds a new sample PAM configuration file for FreeBSD (and
compatible systems). The new sample configuration file is named
campat/pam.conf.freebsd.