Search /usr/pkg/lib/libatomic when linking.
This is for the benefit of NetBSD/powerpc, where we now depends on
the recently added libatomic-links package which leaves symlinks to
the libatomic library files in that directory.  This is so that we
don't automatically pick up other libraries behind the compiler users
back.

Also, match the newly uploaded NetBSD/macppc 9.0 bootstrap (regenerated
with this patch applied to 1.46.0).

No pkgrevision bump, since this is a build fix for NetBSD/powerpc.