Updated devel/ocaml-dune to version 2.7.1.
Changes from the previous version are too many to mention (see the
CHANGELOG.md file in the distribution for full details). From a practical
point of view the package bootstrap procedure has changed now so that
dune-configurator had to be split off as a separate package; there are
also compatibility changes so the package works with newer versions of
OCaml.