Add sysutils/yank.
The yank(1) utility reads input from stdin and display a selection interface
that allows a field to be selected and copied to the clipboard. Fields are
either recognized by a regular expression using the -g option or by splitting
the input on a delimiter sequence using the -d option.

Using the arrow keys will move the selected field. The interface supports
several Emacs and Vi like key bindings, consult the man page for further
reference. Pressing the return key will invoke the yank command and write
the selected field to its stdin. The yank command defaults to xsel(1) but
could be anything that accepts input on stdin. When invoking yank, everything
supplied after the -- option will be used as the yank command.