pkgtools/lintpkgsrc: update to 4.96
Changes since 4.95:

Do not complain about "Bogus: ${_SVN_PKGREVISION:...}" that are
generated by wip/mk/*-package.mk.  Using PKGREVISION to track the
current date of the package is the wrong approach; PKGVERSION_NOREV
should be used instead.  Anyway, this is nothing the average pkgsrc user
should be worried about, so shut up the warning for now.

https://mail-index.netbsd.org/pkgsrc-users/2020/12/16/msg032884.html