Add nancy version 1.0.5
nancy is a tool to check for vulnerabilities in your Golang
dependencies, powered by Sonatype OSS Index, and as well, works
with Nexus IQ Server, allowing you a smooth experience as a Golang
developer, using the best tools in the market!

(No go prefix as I don't think it matters which Go version nancy was
built with. it parses the output from Go itself.)