py-alembic: updated to 1.5.2
1.5.2

[bug] [regression] [versioning]
Fixed regression where new “loop detection” feature introduced in 757 produced false positives for revision names that have overlapping substrings between revision number and down revision and/or dependency, if the downrev/dependency were not in sequence form.

[bug] [environment]
Fixed regression where Alembic would fail to create a transaction properly if the sqlalchemy.engine.Connection were a so-called “branched” connection, that is, one where the .connect() method had been called to create a “sub” connection.


1.5.1

[bug] [commands] [installation]
Fixed installation issue where the “templates” directory was not being installed, preventing commands like “list_templates” and “init” from working.


1.5.0

[changed] [environment]
To accommodate SQLAlchemy 1.4 and 2.0, the migration model now no longer assumes that the SQLAlchemy Connection will autocommit an individual operation. This essentially means that for databases that use non-transactional DDL (pysqlite current driver behavior, MySQL), there is still a BEGIN/COMMIT block that will surround each individual migration. Databases that support transactional DDL should continue to have the same flow, either per migration or per-entire run, depending on the value of the Environment.configure.transaction_per_migration flag.

[changed] [environment]
A CommandError is raised if a sqlalchemy.engine.Engine is passed to the MigrationContext.configure() method instead of a sqlalchemy.engine.Connection object. Previously, this would be a warning only.
[changed]
Alembic 1.5.0 now supports Python 2.7 and Python 3.6 and above, as well as SQLAlchemy 1.3.0 and above. Support is removed for Python 3 versions prior to 3.6 and SQLAlchemy versions prior to the 1.3 series.