math/bc-gh: Update to 3.3.0
## 3.3.0

This is a production release that changes one behavior and fixes documentation
bugs.

The changed behavior is the treatment of `-e` and `-f` when given through
`BC_ENV_ARGS` or `DC_ENV_ARGS`. Now `bc` and `dc` do not exit when those options
(or their equivalents) are given through those environment variables. However,
`bc` and `dc` still exit when they or their equivalents are given on the
command-line.