py-hg-evolve: update to 10.3.0.
This is a feature release that comes with a variety of improvements and
bugfixes. Check the changelog for details. The most notable changes
are: improved rewind logic in cases involving folds, an experimental
ability to perform evolution in-memory (only on Mercurial 5.6 and
newer), improved content-divergence resolution logic in cases involving
parent changes.