hackrf: update to 2021.03.01
This release requires updating your HackRF to the latest firmware.

CPLD bitstreams are now included in and loaded automatically by the firmware.
There is no longer a need to update the CPLD separately.
The clock output (CLKOUT) is now turned off by default.
A new utility, hackrf_clock, provides a way to inspect and modify configuration
of hardware clocks, including turning on and off CLKOUT.
The rad1o badge is now supported by firmware in this release. Firmware from the
HackRF release/repository is different than the original CCCamp firmware and
only supports operation as a USB SDR peripheral.
When using a HackRF One with a PortaPack in "HackRF Mode" (operating as a USB
peripheral), the PortaPack display now indicates operational status and settings.

There have been many more enhancements and bug fixes.
For a full list of changes, see the git log.