tex-cweb{,-doc}: update to 4.2
In the course of the "2021 TeX tuneup", Don Knuth kindly transferred the
responsibility for CWEB to me; see the original CWEB home page
(https://www-cs-faculty.stanford.edu/~knuth/cweb.html), section "CWEB
4.0".

This first new release draws heavily on the multi-decade work collected in
my "CWEBbin" project (https://github.com/ascherer/cwebbin).  I moved
(almost) all stuff from the "*-patch.*" changefiles into the new CWEB
sources and applied all of the "*-ansi.*" changes as well.

Moreover, I made "common.h" the full interface of "common.w" in order to
avoid code redundancy.  And I tried to fix as many of the bugs reported by
email to DEK and/or tex-k at tug.org.

In total, "make; make cautiously; make fullmanual" runs flawlessly with
GCC on Linux and with CLang on MacOS.  And with the help of the remaining
stuff in the "CWEBbin" project, "CWEB" is smoothly integrated in "TeX
Live".