(graphics/R-thematic) import R-thematic-0.1.2
Theme 'ggplot2', 'lattice', and 'base' graphics based on a few
choices, including foreground color, background color, accent color,
and font family. Fonts that aren't available on the system, but are
available via download on 'Google Fonts', can be automatically
downloaded, cached, and registered for use with the 'showtext' and
'ragg' packages.