bash: Add --enable-function-import.
This is required to make 'export -f' work, which is enabled by default on
other OS I checked.  While here sort the list of enabled options so that
it's easier to compare against the list of available options, as there are
still a number that we do not enable, and remove the bogus comment.

Bump PKGREVISION.