net/py-magic-wormhole-transit-relay: import py-magic-wormhole-transit-relay-0.2.1
This repository implements the Magic-Wormhole "Transit Relay", a
server that helps clients establish bulk-data transit connections
even when both are behind NAT boxes. Each side makes a TCP connection
to this server and presents a handshake. Two connections with
identical handshakes are glued together, allowing them to pretend
they have a direct connection.

This server used to be included in the magic-wormhole repository,
but was split out into a separate repo to aid deployment and
development.