go-bin: Hide the signing removal behind DARWIN_CHROOTED.
Unfortunately stripping the signatures breaks running the binaries outside a
chroot according to schmonz, so we're left with no choice but to make this
user-configurable.  It took a lot of effort for me not to name this variable
something far ruder.

This will still mean the go-bin package is ultimately useless when shipped as
part of a package set, but really we only need it for bootstrapping a real
version which should work fine.