Update to 1.1.0. From the changelog:
# Storyboards

Sometimes we might want to see different steps in a workflow or lifetime
of an object. Storyboards are a convenience object to help enable that.

Approvaltests allows us to look at a complete object instead of just
pieces of it. Storyboards allow us to track an object through time.

The mechanism to map time to space that storyboards use is very
analogous to a comic book, but with each frame vertically after each
other so that it works well with the diff tool and shows a progression.