mk/install: fix documentation and speed up install-dirs-from-PLIST
The documentation did not mention that conditional entries in PLIST
files are skipped.  Pkglint doesn't know this either and issues wrong
notes that some directories are redundant in INSTALLATION_DIRS when in
reality they need to be listed there.

While here, filter out some duplicate directories.  This invokes the
command for creating a single directory less often.  Since PLIST files
are usually sorted, files in the same directory tend to be listed near
each other.

The fallback to the plain 'uniq' is needed for SCO_SV, which does not
define TOOLS_PLATFORM.uniq.