devel/R-R.oo: import R-R.oo-1.24.0
Methods and classes for object-oriented programming in R with or
without references.  Large effort has been made on making definition
of methods as simple as possible with a minimum of maintenance for
package developers.  The package has been developed since 2001 and is
now considered very stable.  This is a cross-platform package
implemented in pure R that defines standard S3 classes without any
tricks.