misc/py-anita: update to 2.9.
Changes from 2.8:

Support the new "gpufw" set.

Actually reduce the default test timeout to 1 hour as documented in
version 2.5 when anita is invoked as a program rather than as a
library.

Log a message to the structured log file on timeout so that the time
when the timeout occurred can be determined from the message timestamp.

Define a scratch disk for alpha (wd1c) so that ATF test results can
be exported to the host.

In addition to the plain-text ATF test report printed to the console
as the tests progress, also store a clean copy without other console
output in test.txt and export it to the host.  Suggested by
Greg Troxel.

Fix failure to destroy the Xen DomU used by the installation stage
before trying to create the one for the boot stage when running
under Python 3.

Fix failure to log in after reboot when running "anita test" on a
self-resizing image without running a separate "anita install" first.
From Joerg Sonnenberger.

Support coypu's proposed sysinst entropy patch.

Remove extra newline in reponse to entropy prompt.