textproc/hgrep: update to 0.1.9
-Support multiple regions highlighting. In v0.1.8, matched region highlighting
was added but it only highlighted the first match in the line. Now all matched
regions are highlighted. Note that region highlighting is available when using
hgrep in combination of syntect-printer and ripgrep features.
-Add --ascii-lines flag for terminals which does not support rendering unicode
characters well. With this flag, unicode characters like '│' or '─' are replaced
with ASCII characters '|' or '-'. This feature is only supported by
syntect-printer (use -p syntect).
-Add --first-only (-f) flag to show only the first snippet per file. This is
useful when you want to look around the results.
-Fallback to a minimap border color when no gutter background color is found.
-Reduce number of redundant color codes output to stdout by about 21.5% in test
cases. This also improves performance by about 6%. See the commit for details.
-Performance of built-in grep was improved 20~80% when there are so many files
to search. Previously the implementation collected all paths before searching a
pattern in them, but with this improvement, the paths are now streamed.
See the commit for details.
-Use mimalloc as global allocator for better performance. This improves
performance by 0~55% in our benchmarks. See the commit for details.
-(Dev) Running unit tests is about 8.5x faster by caching assets for syntax
highlighting.