fonts/UFAS-otf: 20200320 aka version 13.00, successor to fonts/TextFonts
The original TextFonts distfile is no longer available.
The home page has moved (as linked from the previous version) and
reorganized, and now it lists several more/different related fonts. It
would be silly not to package them too, since they fit within the
original package description.  The font files are now OTF instead of
TTF. The differences are so much that a new package with a new
appropriate name seemed better than keeping the old package name.

No changelog was found.