shells/starship: update to 1.0.0
This is it! Starship v1.0.0!
Truth be told, there's no real cause for celebration. After 100 releases of
Starship, we felt it was about time we begin sticking to proper SemVer.

No breaking changes to worry about in this one – we just wanted to let you rest
easy knowing that future releases won't have breaking changes without a major
version bump.