py-httpcore: updated to 0.14.1
0.14.1 (November 12th, 2021)

max_connections becomes optional.
certifi is now included in the install dependancies.
h2 is now strictly optional.


0.14.0 (November 11th, 2021)

The 0.14 release is a complete reworking of httpcore, comprehensively addressing some underlying issues in the connection pooling, as well as substantially redesigning the API to be more user friendly.

Some of the lower-level API design also makes the components more easily testable in isolation, and the package now has 100% test coverage.

There's some other neat bits in there too, such as the "trace" extension, which gives a hook into inspecting the internal events that occur during the request/response cycle. This extension is needed for the HTTPX cli, in order to...

Log the point at which the connection is established, and the IP/port on which it is made.
Determine if the outgoing request should log as HTTP/1.1 or HTTP/2, rather than having to assume it's HTTP/2 if the --http2 flag was passed. (Which may not actually be true.)
Log SSL version info / certificate info.
Note that curio support is not currently available in 0.14.0. If you're using httpcore with curio please get in touch, so we can assess if we ought to prioritize it as a feature or not.