mold: update to 1.0.0.
mold 1.0 is the first stable and production-ready release of the high-speed
linker. On Linux-based systems, it should "just work" as a faster drop-in
replacement for the default GNU linker for most user-land programs. If you
are building a large executable which takes a long time to link, mold is
worth a try to see if it can shorten your build time. mold is easy to build
and easy to use. For more details, see README.

mold is created by a person who knows very well as to how the Unix linker
should behave, as I'm also the original creator of the current version of
the LLVM lld linker.

There's no fancy new features in 1.0. Actually, 1.0 is very similar to
0.9.6. That being said, we'd like to make it clear by incrementing a major
version number that mold for Linux is now stable.

Changes since mold 0.9.6:

- -start-lib and -end-lib options are added for compatibility with GNU
  gold and LLVM lld.
- More ARM64 relocations are supported.
- Compatibility with glibc 2.2 or prior has improved. (#120)
- Compatibility with valgrind has improved. (#118)
- -Bno-symbolic option has been supported.
- -require-defined option has been supported.