devel/ruby-activesupport70: add package version 7.0.0
Active Support -- Utility classes and Ruby extensions from Rails

Active Support is a collection of utility classes and standard library
extensions that were found useful for the Rails framework.  These additions
reside in this package so they can be loaded as needed in Ruby projects
outside of Rails.

This is for Ruby on Rails 7.0.