py-supervisor: updated to 4.2.4
4.2.4 (2021-12-30)

Fixed a bug where the --identifier command line argument was ignored. It was broken since at least 3.0a7 (released in 2009) and probably earlier. Patch by Julien Le Cléach.

4.2.3 (2021-12-27)

Fixed a race condition where an rpcinterface extension that subscribed to events would not see the correct process state if it accessed the the state attribute on a Subprocess instance immediately in the event callback. Patch by Chao Wang.
Added the setuptools package to the list of dependencies in setup.py because it is a runtime dependency. Patch by Louis Sautier.
The web interface will now return a 404 Not Found response if a log file is missing. Previously, it would return 410 Gone. It was changed because 410 is intended to mean that the condition is likely to be permanent. A log file missing is usually temporary, e.g. a process that was never started will not have a log file but will have one as soon as it is started.