Add sleepwatcher, a daemon that runs a Unix command on various macOS sleep or wake events.
SleepWatcher is a daemon for macOS that monitors sleep, wakeup and
idleness of a Mac. It can be used to execute a Unix command when the Mac
or the display of the Mac goes to sleep mode or wakes up, after a given
time without user interaction or when the user resumes activity after a
break or when the power supply of a Mac notebook is attached or
detached. It also can send the Mac to sleep mode or retrieve the time
since last user activity.