apr-util: Disable memset_s on SunOS.
The configure test correctly defines __STDC_WANT_LIB_EXT1__=1 as is
required when using the _s set of functions, so memset_s is found, but
then the actual source code does not, so you're left with undefined
references.

Disabling completely is simpler than having to rearrange everything
else to work, and will fall back to portable features.