sysutils/hs-hslogger: import hs-hslogger-1.3.1.0
hslogger is a logging framework for Haskell, roughly similar to Python's
logging module.

hslogger lets each log message have a priority and source be associated
with it. The programmer can then define global handlers that route or
filter messages based on the priority and source. hslogger also has a
Syslog handler built in.