lang/guile30; Add version 3.0.7
Packaged by gdt and myself in wip.

Compared to guile 2.2, this is a major release which implements a JIT
compiler.  See NEWS in the sources for more details.


GUILE, GNU's Ubiquitous Intelligent Language for Extension, is a library
that implements the Scheme language plus various convenient facilities.
It's designed so that you can link it into an application or utility to
make it extensible. Our plan is to link this library into all GNU programs
that call for extensibility.

This is guile 3.0, the current stable version.