vte3: Update to 0.68.0.
The recent vala update broke this build, this seems to get it back up
and running again.  Unfortunately upstream have switched to a "git log"
changelog which is only available in distfiles and is all but useless,
so no list of useful changes since 0.60.3 other than "bugfixes?".

Tested on NetBSD and SmartOS.