tex-readarray{,-doc}: update to 3.1
Version 3.0 of readarray introduces several new features.  Arrays can be
initialized with the \initarray command, rather than being read from a
file.  With \setvalue, the tokens of individual array cells can be revised
on demand.  The contents of one array can be merged into another by way of
the \mergearray command.  The \typesetarray command allows for versatile
ways to format your array output, including by way of tabular
environments. The way end-of-line characters can be handled on a data-read
is made more versatile with this package version.