phpldapadmin: package version 1.2.6.3
This changes the upstream to leenooks/phpLDAPadmin on GitHub.

This also includes a patch from Debian at
https://packages.debian.org/source/sid/phpldapadmin
(phpldapadmin_1.2.6.3-0.2.debian.tar.xz) in order to support newer
versions of PHP. This package had become unusable with the version
of PHP shipped by pkgsrc by default (7.4 as of today).

Tested on NetBSD/amd64, Darwin/amd64.