mysql57: updated to 5.7.38
Changes in MySQL 5.7.38

SQL Function and Operator Notes

When the mysql client was started with --default-character-set=utf8mb4, successive calls to the UUID() function returned duplicate values.

Security Notes

The linked OpenSSL library for MySQL Server has been updated to version 1.1.1n from 1.1.1l. Issues fixed in OpenSSL are described at https://www.openssl.org/news/cl111.txt and at http://www.openssl.org/news/vulnerabilities.html.

Functionality Added or Changed

The default for the group_replication_transaction_size_limit system variable, which sets the maximum transaction size that a replication group accepts, is changed from zero (no limit) to 150000000 bytes (approximately 143 MB), which is the same as the default in MySQL 8.0. Setting a limit for this system variable by default helps to avoid delays or errors caused by excessively large transactions. Transactions above the limit are rolled back and are not sent to Group Replication's Group Communication System (GCS) for distribution to the group. If your Group Replication servers previously accepted transactions larger than the new default limit, and you were allowing group_replication_transaction_size_limit to default to the old zero limit, those transactions will start to fail after the upgrade to the new default. You must either specify an appropriate size limit that allows the maximum message size you need the group to tolerate (which is the recommended solution), or specify a zero setting to restore the previous behavior.

The myisam_repair_threads system variable and myisamchk --parallel-recover option are deprecated; expect support for both to be removed in a future release of MySQL.

Values other than 1 (the default) for myisam_repair_threads produce a warning.

Bugs Fixed

InnoDB: A missing null pointer check for an index instance caused a failure.

InnoDB: Purge threads processed undo records of an encrypted table for which the tablespace was not loaded, causing a failure.

InnoDB: Incorrect AUTO_INCREMENT values were generated when the maximum integer column value was exceeded. The error was due to the maximum column value not being considered. The previous valid AUTO_INCREMENT value should have been returned in this case, causing a duplicate key error.

Partitioning: In some cases, establishing a connection to MySQL server could fail if the .ibd file for a partition was missing.

Statements that cannot be parsed (due, for example, to syntax errors) are no longer written to the slow query log.

It was not possible to revoke the DROP privilege on the Performance Schema.

A page cleaner thread timed out as it waited for an exclusive lock on an index page held by a full-text index creation operation on a large table.

A memory leak occurred if mysqldump was used on more than one table with the --order-by-primary option. The memory allocated for sorting each table’s rows is now freed after every table, rather than only once.

mysqld_safe log message textual errors were corrected.