repmgr: updated to 5.3.2
repmgr 5.3.2 is a minor release

Bug fixes

repmgr node status: fix output with --downstream --nagios option combination.

repmgr standby clone: don't treat inability to determine the cluster size as a fatal error.

The cluster size is displayed for informational purposes and is not essential for execution of the clone operation. As the repmgr user may not have permissions for all databases in the cluster, ignore the cluster size query if it fails.

repmgrd: ensure the witness node record on the primary is always marked as active if previously marked inactive.

repmgrd: if standby_disconnect_on_failover is set, verify repmgr is a superuser before attempting to disable the WAL receiver.

If the repmgr user is a non-superuser, and a replication-only user exists, ensure redundant replication slots are dropped correctly even if the -S/--superuser option is not provided.