Build fixes for devel/ocaml-bigstringaf. Revbump.
The earlier fixes by mef@ (using make) don't solve the problem properly,
since the Makefile only calls dune anyway. However, it does so with
different parameters, most notably not using the -j parameter (leaving
dune to autodetect the number of jobs). That (and the fact that I can't
reproduce the problem) suggests to me that parallel building is the culpit.
I've disabled it for now and will file a bug upstream.