labltk: update to 8.06.12.
Fix some pkglint.

Changes:

LablTk 8.06.12 with OCamlBrowser

Compatibility release to make OCamlBrowser work with OCaml 4.14 (and only this version)

LablTk 8.06.11 with OCamlBrowser

This is a bug fix release of LablTk and OCamlBrowser.
It also makes OCamlBrowser work on OCaml 4.13 (and only this version).

    Disable -warn-error for users, add all-devel target for developers
    Add ommitted labels
    Update ocamlbrowser for ocaml 4.13

LablTk 8.06.10 with OCamlBrowser

This is a bug fix release of LablTk and OCamlBrowser.
It also makes OCamlBrowser work on OCaml 4.12 (and only this version).

    Fix bugs in ocamlbrowser that prevented exploring types in source files
    Install camltkwrap.{cmi,cmx} (report by Pascal Raymond)

LablTk 8.06.9 with OCamlBrowser 4.11.0

New release of LablTk and OCamlBrowser

    LablTk is compatible with ocaml 4.07 and newer
    OCamlBrowser is compatible with ocaml 4.11

Fixes a configuration error with clang 12 (observed on MacOS).

LablTk 8.06.8 with OCamlBrowser 4.10.0

New release of LablTk and OCamlBrowser

    LablTk is compatible with ocaml 4.07 and newer
    OCamlBrowser is compatible with ocaml 4.10

Fixes a number of bugs in OCamlBrowser, particularly concerning the file selection window.

LablTk 8.06.7

New release fixing compatibility problems for ocaml 4.09.

    still compatible with ocaml 4.07
    ocamlbrowser compatible with ocaml 4.08 and 4.09

opam packages should be labltk.8.06.7 and ocamlbrowser.4.08.1

LablTk 8.06.6

New release for ocaml 4.08.

    compatible with ocaml 4.08.0 (required for ocamlbrowser)
    compiles and installs without ocamlbrowser (use make library opt rather than make all opt).

Using opam, two packages are provided: labltk.8.06.6 and ocamlbrowser.4.08.0