graphics/ocaml-cairo: Remove gtk and pango from default options
The gtk and pango options depend on ocaml-lablgtk (which is gtk2), but
ocaml-lablgtk3 depends on ocaml-cairo.  For now, don't suggest the
options, as the only pkgsrc user is ocaml-lablgtk3.

Discussed with $MAINTAINER long ago.  Perhaps these need a split
package, and perhaps packages that depend on ocaml-cairo might need
updating -- but gtk2 is obsolete and those packages should be moving
(perhaps upstream) to lablgtk3.

(Committed separately from 0.6.3 update for ease of bisecting and
reverting if necessary.)