libX11: update to 1.8.1.
1.8.1

This release fixes the --enable-thread-safety-constructor option to the
configure script to work as intended.  In the previous release, the changes
for this option may not have been enabled when the option was not specified
or when the --enable option was specified.

While we have enabled it by default, believing that doing so will reduce
the number of bugs users encounter running libX11 clients, in some cases
it may expose bugs in which clients had previously gotten away with calling
libX11 functions while a libX11 lock is already held, and thus now deadlock,
as discussed in https://gitlab.freedesktop.org/xorg/lib/libx11/-/issues/157 .

Alan Coopersmith (2):
      configure.ac: Fix --enable-thread-safety-constructor
      libX11 1.8.1

1.8

The highlight of this release is that we now try to initialize thread
safety ourselves, rather than hope the application does it. This should
resolve a number of long-standing bugs with the libxcb integration,
since the socket handoff mechanism essentially has to be thread-safe.
But regressions are always possible, so please report any issues
introduced by this release.

Adam Jackson (2):
      global: call XInitThreads() from the library's constructor
      libX11 1.8

Alan Coopersmith (4):
      XkbUpdateKeyTypeVirtualMods: always initialize mask
      XKBgeom.h: Remove XKB_IN_SERVER section
      XKBgeom.h: Add _Xconst qualifier to char * arguments in functions
      XkbOpenDisplay: Add _Xconst qualifier to display name argument

Benno Schulenberg (7):
      nls: remove redundant Khmer digraphs, already defined in included file
      nls: remove misplaced sequences from Ethiopian and Greek compose files
      nls: remove eight pointless self-producing compose sequences
      nls: add Multi-key sequences for abovedot, diaeresis, and ogonek
      nls: make the Ethiopian compose sequences actually work
      lcCT: use the correct index for checking the second byte
      docs: add release notes for versions 1.7.3, 1.7.4, and 1.7.5