license.mk: add Beer-Ware to DEFAULT_ACCEPTABLE_LICENSES
As recently discussed on tech-pkg, this should be considered "Free or
Open Source", so add it accordingly. Probably this was never done
before just because it is so uncommon.

Debian allows inclusion of it as such (e.g., it's one of the
attributions they note applies to OpenSSH code in their packaging of
it), as do other mainstream Linux distros like Fedora.

Debian package review where this is stated:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=685575

Fedora policy:
https://fedoraproject.org/wiki/Licensing/Beerware