libX11: report the thread safety constructor build state (NFC)
1.8 added a feature (--enable-thread-safety-constructor) which is
enabled by default and can cause issues like hangs with X11 clients
that aren't coded properly. (This affects more than one DE in pkgsrc.)
When the feature was added, it didn't include any output from configure
itself that reports what the state of the option is set to.

No change to package defaults or resulting binaries, this is just for
informational purposes for anyone else dealing with this, should they
want to disable the setting, or just recognize what state the build
used.