freeradius: correct config installation and de-installation
The script that un-installs the configuration content placed under the
PKG_SYSCONFDIR wasn't removing one of the files that's installed. (This
had been incorrect since it was added to the installation script back
in 2017.)

Separately, simplify the previous changes I made to fix installation of
configuration files to PKG_SYSCONFDIR. I had used redundancy to force
certain directories necessary for startup to be present and populated,
but the pre-existing code in the scripts had almost all of it covered,
just not the creation of the directories themselves and a single file.