mk: Add OPSYS_EXPLICIT_LIBDEPS and enable on SunOS.
Most systems use GNU ld, which will happily pull in symbols required by a
program even if they are only available via implicit library dependencies.  The
SunOS linker is stricter, and if a program uses a symbol then the library that
defines that symbol must be an explicit dependency.

This mostly causes problems with libiconv and libintl, both of which Linux
bundles in its C library, so a lot of third-party software does not correctly
check for them.  Until now we've had to add many, many overrides, along with
variables such as BROKEN_GETTEXT_DETECTION which nowadays only has limited
effectiveness.

The situation appears to be getting worse, especially with software built with
meson, and so both libiconv and gettext-lib will now automatically add the
correct LDFLAGS if the OPSYS sets OPSYS_EXPLICIT_LIBDEPS=yes.

This isn't perfect.  For one it isn't really an OPSYS setting as you can try to
use GNU ld on SunOS, it just doesn't work very well.  It should also really be
done via the wrappers rather than exposing LDFLAGS, but we do not yet have an
approved patch for doing this.  However it does improve the current situation.