misc/wthrr: import package
When spending a lot of time in the CLI it can be handy to have a little
companion in there who knows about the weather.

That's where wthrr comes in. She lives in your terminal and her passion is
the weather.

Even when you don't spend a lot of time in the terminal. She has a sunny
personality, so you can visit her once in a while.