Switch desktopdb.mk to not depend on desktop-file-utils
Consumers of desktop entries still register a full dependency
on desktop-file-utils, but apps that install a .desktop file
in share/applications do not. They still have a POST-INSTALL
step to call UPDATE_DESKTOPDB, but only if it is already
installed.

Should be not functional difference for any system with a
consumer of desktop entries installed, but for an app
with a .desktop entry the build of python, meson & glib2
may be deferred. Helpful for small apps and lower power
systems