chat/matrix-synapse: MESSAGE exorcism
This had MESSAGE to given an example config generation command (which
probably follows upstream docs), and to warn about the deficient
sqlite in NetBSD base.  Move them to files/ and install as
documentation.  Add a caution to the Makefile near the sqlite include.

(It remains to actually fix the sqlite issue.)