py-jupyter_core: updated to 5.0.0
5.0.0

Major Changes

Prefer Environment Level Configuration

We now make the assumption that if we are running in a virtual environment, we should prioritize the environment-level `sys.prefix` over the user-level paths. Users can opt out of this behavior by setting `JUPYTER_PREFER_ENV_PATH`, which takes precedence over our autodetection.

Migrate to Standard Platform Directories

In version 5, we introduce a `JUPYTER_PLATFORM_DIRS` environment variable to opt in to using more appropriate platform-specific directories.  We raise a deprecation warning if the variable is not set.  In version 6,  `JUPYTER_PLATFORM_DIRS` will be opt-out.  In version 7, we will remove the environment variable checks and old directory logic.

Drop Support for Python 3.7

We are dropping support for Python 3.7 ahead of its official end of life, to reduce maintenance burden as we add support for Python 3.11.

Enhancements made

- Use platformdirs for path locations
- Try to detect if we are in a virtual environment and change path precedence accordingly

Bugs fixed

- Add current working directory as first config path
- Fix inclusion of jupyter file and check in CI

Maintenance and upkeep improvements

- Bump github/codeql-action from 1 to 2
- Bump actions/checkout from 2 to 3
- Add dependabot
- Adopt jupyter releaser
- Add more typing
- Require Python 3.8+
- Use hatch backend
- Switch to flit build backend
- is_hidden: Use normalized paths

Documentation

- Update broken link to `Contributing` guide