munge TTP and IPTM into aliases
Some time between 2011 and 2013, IANA assigned protocol 84 to IPTM even
though it had already assigned it to TTP.  This was previously merged
manually (e.g., in rev1.25, and rev 1.28), but required the person
regenerating the data to remember to do this.  This change here now
makes that change automatic, ugly as it is.