py-html-sanitizer: updated to 1.9.3
1.9 (2020-01-20)

Added Python 3.8 to the CI matrix.
Be able to keep the <style> tag by adding it to tags.
Added a style check to the CI matrix.

1.8 (2019-11-21)

Actually added support for customizing lxml's autolinking behavior using a dictionary argument.
Stopped removing explicitly allowed attributes.
Removed id from allowed attributes of <a> tags to provide an additional layer of defense against DOM clobbering attacks.
Added an element preprocessor which assigns the id value to the name attribute of anchors if name isn't set or empty. This should provide additional backwards compatibility making the id removal less of a problem when using named anchors.

1.7 (2019-02-19)

Added a system check which validates sanitizer configurations early when using Django.
Fixed an edge case where passing in an empty allowed tags list would unexpectedly and silently not remove any tags at all (because that's the way lxml's cleaner works).
Changed the sanitizer tags, empty and separate options to also accept any iterable, not just sets.
Changed the lru_cache import in the Django module to try functools first.
Fixed the tag merging to also check tags in empty. This means that e.g. consecutive <hr> tags are also merged now when using the default settings.
Made it possible to override the set of tags processed as whitespace. The default set is {"br"} which preserves the current behavior of stripping breaks from the beginning or end of tags' content.