libXxf86dga: update to 1.1.6.
The primary change in this release is dropping support for a number of
OS'es from the previous millenium.  If you're still running an OS that
has not had a new release in the past quarter of a century, then you
probably don't want to try installing this new release on it.

Alan Coopersmith (9):
      Build xz tarballs instead of bzip2
      Fix spelling/wording issues
      gitlab CI: add a basic build test
      unifdef __UNIXOS2__
      Remove SunOS 4 support (defined(sun) && (SVR4))
      unifdef ISC
      unifdef Lynx
      unifdef DGUX
      libXxf86dga 1.1.6