mail/mm: import package from wip.
Columbia MM (Mail Manager) is an interactive, text-mode mail client for
UNIX, based on the TOPS-20 MM program. It is written in C using the CCMD
(TOPS-20 COMND Jsys in C) package developed at Columbia University.

Project is currently hosted and maintained on Kermit Project's FTP
server.

CCMD libraries are built alongside MM and included in the package.

This package uses a stripped-down tarball with all the required files
(+ a couple of fixes) hosted on my account at ftp.NetBSD.org, for
reasons of conveniency.
Interesting docs and historical notes are included too.