i3: update to 4.22.
The biggest change in this release is the merge of the i3-gaps fork.
The i3-gaps fork was the most popular fork of i3, adding the option to
show gaps between tiled windows and/or the screen edges.

See https://i3wm.org/docs/userguide.html#gaps for more details.

Instead of maintaining two versions of i3 (both upstream and downstream,
meaning in Linux distributions and other package collections),
we concluded it would be better for everyone to merge this feature.

For users of i3: gaps are off by default, so there is no change in behavior.
For users of i3-gaps: the configuration is compatible, so you can switch
to i3 v4.22 or newer, without any changes in behavior.

Thanks to Ingo Bürk for maintaining i3-gaps for many years,
for becoming a core i3 maintainer and for helping make this merge possible!