Import net/gnunet as gnunet version 0.19.2
Imported from wip/gnunet

GNUnet is a framework for secure peer-to-peer networking that does not
use any centralized or otherwise trusted services.

A first service implemented on top of the networking layer allows
anonymous censorship-resistant file-sharing.

GNUnet uses a simple, excess-based economic model to allocate
resources.

Peers in GNUnet monitor each others behavior with respect to resource
usage; peers that contribute to the network are rewarded with better
service.