mail/thunderbird: Update to version 102.6.1.
This is copied from wip/thunderbird, which builds with
more recent versions of rust.

This is the version just before this change was implemented:

https://blog.thunderbird.net/2023/01/important-message-for-microsoft-office-365-enterprise-users/

and this version will stay here for a while before we update
again.

Call me lazy, but I don't think it's worth anyone's time to paste
in the change log from version 78 to 102.6.1 here, and it doesn't
appear to be readily accessible in the source distribution.  That
said, the list of security fixes implemented between version 78.12
and 102.6.1 is available at

https://www.mozilla.org/en-US/security/known-vulnerabilities/thunderbird/#thunderbird102.6.1

and the release notes are available at

https://www.thunderbird.net/en-US/thunderbird/releases/