xorg-cf-files: update to 1.0.8.
This package contains the configuration files, platform data, & Makefile
templates for the archaic Imake build tool.  This is a minor update to
ship the contributed support for Linux on RISC-V & LoongArch, and
AArch64 on Darwin (MacOS), as well as to fix incompatibility when using
GNU binutils 2.36 and later on FreeBSD, and to correct the setting of
the rarely (never?) used ENCODINGSDIR in site.def.

As before, while we continue to accept patches and publish updates for the
benefit of other packages which still use Imake to build, we have not been
actively maintaining, testing, or adding new OS or platform support for
Imake since 2005.  X.Org strongly encourages anyone still using Imake to
move to meson or another modern/maintained build tool.

For example, this release adds support for installing with meson as well
as the traditional GNU autoconf.  Please test building with meson and
report any issues you hit to our gitlab issue tracker at:
https://gitlab.freedesktop.org/xorg/util/cf/-/issues
as we plan on removing the autoconf build support in a future release.

Alan Coopersmith (12):
      gitlab CI: add a basic build test
      Build xz tarballs instead of bzip2
      Fix spelling/wording issues
      Use AC_CONFIG_FILES to replace the deprecated AC_OUTPUT with parameters
      gitlab CI: stop requiring Signed-off-by in commits
      More typo fixes
      configure.ac: Don't pass through '${FONTROOTDIR}' into site.def
      meson: Add option to build with meson
      meson: modernize some usage in meson.build & .gitlab-CI.yml
      meson: simplify installation of empty *.def files
      configure: Switch to XORG_DEFAULT_NOCODE_OPTIONS
      xorg-cf-files 1.0.8

Emil Velikov (1):
      autogen.sh: use quoted string variables

Marek Szuba (1):
      Linux/RISC-V support

Peter Hutterer (1):
      autogen.sh: use exec instead of waiting for configure to finish

WANG Xuerui (1):
      Add support for Linux/LoongArch

Weijia (1):
      Add support for aarch64-darwin

Yasuhiro Kimura (1):
      Fix incompatibility issue of GNU binutils 2.36 and later on FreeBSD.