textproc/ruby-unicode-scripts: add package version 1.8.0
This package is required by new version of ruby-kramdown-rfc2629.


Unicode::Scripts

Retrieve the Unicode script(s) a string belongs to.  Can also return the
Script_Extension property which is defined as characters which are "commonly
used with more than one script, but with a limited number of scripts".

Unicode version: 15.0.0 (September 2022)