x11/ocaml-lablgtk: Disable gtksourceview2 option
For reasons that are too hard to figure out, building with this option
enabled now results in a build failure.  Most depending packages have
migrated to ocaml-lablgtk3, and none are known to depend on this
option.  Remove the option from SUGGESTED.  It is left for future work
to anyone that cares to debug this, but probably there is no such
person.

ok jaapb@