sqlite3: updated to 3.41.1
version 3.41.1 (2023-03-10):

Provide compile-time options -DHAVE_LOG2=0 and -DHAVE_LOG10=0 to enable SQLite to be compiled on systems that omit the standard library functions log2() and log10(), repectively.
Ensure that the datatype for column t1.x in "CREATE TABLE t1 AS SELECT CAST(7 AS INT) AS x;" continues to be INT and is not NUM, for historical compatibility.
Enhance PRAGMA integrity_check to detect when extra bytes appear at the end of an index record.
Fix various obscure bugs reported by the user community. See the timeline of changes for details.