digikam: update to 7.10.0.
After three months of active maintenance and other bugs triage,
the digiKam team is proud to present version 7.10.0 of its open
source digital photo manager.

This new version arrives with 50 triaged and closed files in bugzilla.

See below a list of major improvements included in this release:

* Enable copying file path to the clipboard in Windows.
* Improve support of DXO tags.
* Add support for the new online documentation.
* Fix remote Mariadb database connection under macOS.
* Fix the Cancel button in the Import Tool.
* Fix crashes with last OpenCV 4.7.0.
* Improve internal digiKam trash behavior.
* Fix similarity search settings management.

The application internationalization has also been updated. digiKam
and Showfoto are proposed with 57 different languages for the
graphical interface. Go to Settings/Configure Languages dialog and
change the localization as you want. Applications need to be
restarted to apply changes. If you want to contribute to the
internationalization of digiKam, please contact the translator
teams, following the translation how-to. The statistics about
translation states are available here.