lang/ruby27: update to 2.7.8
Ruby 2.7.8 Released			Posted by usa on 30 Mar 2023

Ruby 2.7.8 has been released.

This release includes security fixes. Please check the topics below for
details.

* CVE-2023-28755: ReDoS vulnerability in URI
* CVE-2023-28756: ReDoS vulnerability in Time

This release also includes some build problem fixes. See the GitHub releases
for further details.

After this release, Ruby 2.7 reaches EOL.  In other words, this is expected
to be the last release of Ruby 2.7 series.  We will not release Ruby 2.7.9
even if a security vulnerability is found (but could release if a severe
regression is found).  We recommend all Ruby 2.7 users to start migration to
Ruby 3.2, 3.1, or 3.0 immediately.