tex-currfile{,-doc}: add version 0.8
The package provides macros holding file name information
(directory, base name, extension, full name and full path) for
files read by LaTeX \input and \include macros; it uses the
file hooks provided by the author's filehook. In particular, it
restores the parent file name after the trailing \clearpage of
an \included file; as a result, the macros may be usefully
employed in the page header and footer of the last printed page
of such a file. The depth of inclusion is made available,
together with the "parent" (including file) and "parents" (all
including files to the root of the tree). The package
supersedes FiNK.