tex-mpchess{,-doc}: add version 0.4
This package allows you to draw chess boards and positions. The
appearance of the drawings is modern and largely inspired by
what is offered by the excellent web site Lichess.org. Relying
on MetaPost probably allows more graphic flexibility than the
excellent LaTeX packages. This package is in beta version, do
not hesitate to report bugs, as well as requests for
improvement