libnitrokey: import as security/libnitrokey version 3.8
libnitrokey is a project to communicate with Nitrokey Pro and Storage devices
in a clean and easy manner. Written in C++14, testable with py.test and Catch
frameworks, with C API, Python access (through CFFI and C API, in future with
Pybind11).

The development of this project is aimed to make it itself a living
documentation of communication protocol between host and the Nitrokey stick
devices.