mksh: Remove defer-always flag from printf.
Previously this deferred to a printf binary if one was found in $PATH, which
completely defeats the point of using the builtin for performance reasons.  We
always create a printf wrapper so one was always found.  A nasty side-effect of
this was that with TOOLS_PLATFORM.printf=printf we would end up in a recursive
loop.

Now printf will always use the builtin.  Bump PKGREVISION.