print/poppler: Stop optionally calling gperf.
This is apparently unnecessary, and the behaviour of a package build
should not depend on what other packages happen to be lying around
installed at the same time without an explicit dependency.

Perhaps we might prefer to always use gperf instead of using the
pregenerated files; in that case this needs TOOL_DEPENDS and the path
must be set to ${TOOLBASE}/bin/gperf.