shells/nushell: update to 0.82.0
Today, we're releasing version 0.82 of Nu. This release focuses on polishing
the user experience.

Note: Beginning with this release, Nushell is switching to a four-week release
schedule.
For details see, https://www.nushell.sh/blog/2023-06-27-road-to-1_0.html

NOTE: The optional dataframe functionality is available
using --features=dataframe

This feature is not enabled by default on pkgsrc.
To compile with this feature add 'dataframe' to the list of features defined
(CARGO_FEATURES+=) in the package Makefile.

As part of this release, we also publish a set of optional plugins.

Themes of this release/New features:
 - Custom datetime format (@fnordpig)
Until now, Nushell did not allow changing the default format of dates and times,
leading quite often to undesired rendering, e.g. seeing 'A year ago' in the
output of ls when one wants to know the exact time.
#9500 introduces a new config option to change the default output of dates and
times.

 - Support for comments in multiline pipes (@WindSoilder)
From now on, either in the REPL, in scripts or in modules, you can comment out
any line of a multiline pipe without breaking its execution. This hopefully
makes debugging and testing part of a pipeline much easier!

 - Continuing cratification effort
On the road to 1.0, we have started moving commands around to make place for
those which really matter. As this process involves the creation of some new
Rust crates, e.g. nu-cmd-extra, we call it cratification 😃
With this new release, the bytes commands have been moved behind the extra
feature (use cargo build --features extra to activate them back).
Apart from moving less stable or common ocmmands behind flag, this should also
improve compile times for everyone.

 - Splitting the $nu variable (@jntrnr)
Nushell 0.82 introduces a new familly of stat commands, the scope ... commands.
These commands allow you to access some information about the current Nushell
instance that were previously available in the $nu variable!
e.g. $nu.scope.commands now becomes scope commands.
On the other hand, $nu becomes much smaller and only holds simple paths and
constants.

 - Entering Linguist support (@hustcer)
With the efforts of @hustcer, Nushell is now an official language of Linguist,
allowing for better language support in GitHub.

 - REPL: Mark line entries as excluded from the history by leading with a space
 (samlich)
In some situations you may want to avoid storing an entry inside your shell
history. For example if it includes a secret or running it again in a different
context could be dangerous. To stop nushell from storing a particular line in
its history you can now start with a space and it will be ignored. You can
still recall your last entry once by pressing the Up-Arrow but the command will
never be written to disk and disappear as soon as you submit another command.

 - New commands
     http options
     scope aliases
     scope commands
     scope engine-stats
     scope modules
     scope variables
     scope

 - Breaking changes
    #9509 Cratification moves the bytes commands to nu-cmd-extra. They are thus
     only accessible with --features extra
    #9496 Clean up config by removing legacy options that were previously
     deprecated.
    #9487 Split $nu variable into scope commands and simpler $nu
    #9427 The filesize prefixes Z/Zi have been removed as their underlying
     values where not correctly representable.
    #9407 In the ls --long output the uid column has been renamed to user to be
     consistent with group
    #9368 Don't allow save command to save both stdout and stderr to the same
     file