bootstrap: Use /usr/bin/bash more on SunOS
Document the meanings of bootstrap_foo vs fooprog.

On SunOS, if we find /usr/bin/bash and use it for bootstrap_sh, also
use it for shprog.  Resolves failure to run libarchive's configure on
old SunOS.  (It seems likely libarchive's configure is using modern
constructs that are best avoided, but it is also good to use a better
shell if available.)

Problem and solution from Vivianne Ombretta via pkgsrc-users.  The
explanatory comments are gdt's fault.