sysutils/hw-probe: import package from wip.
Originally packaged by author Andrey Ponomarenko
<andrew.v.ponomarenko@gmail.com>. Revised and updated by @vins.


hw-probe is a tool to probe for hardware, check operability and find
drivers with the help of BSD hardware database.

Probe is a snapshot of your computer hardware state and system
logs.  The tool checks operability of devices by analysis of logs
and returns a permanent URL to view the probe of the computer.