py-atpublic: updated to 4.0
4.0 (2023-06-05)
Drop Python 3.7 support
Remove public.install() which was used to inject the public and private functions into the builtins namespace. This isn’t very helpful and could be actively harmful. Explicit is better than implicit.
The functional form of public() now returns the argument values in the order they are given. This allows you to explicitly bind those values to names in the global namespace. While this is redundant, it does solve some linter problems.
Switch from flake8 and isort to ruff for code quality.
Bump dependencies.