mk: make NO_BIN_ON_* warning messages agnostic of technology
Replace references to "CD-ROM" and "FTP" with "physical media" and
"public networks".

Use the word "shared" instead of "placed". No license prevents the
user from placing binary packages on physical media for personal use.

If NO_BIN_ON_CDROM and NO_BIN_ON_FTP are both set and the reason is
the same, as is the case for most restricted packages, then only warn
once during the package phase to cut down noise.