games/nethack-lib: updated NetHack to version 3.6.7.
NetHack 3.6.7 is the official release of NetHack that follows NetHack 3.6.6.

This release primarily corrects a security issue present in NetHack versions
3.6.2, 3.6.3, 3.6.4, 3.6.5, and 3.6.6. For details about the security issues
please see https://www.nethack.org/security. You are encouraged to update to
NetHack 3.6.7 as soon as possible.

Here's a brief synopsis of the handful of bug fixes included in NetHack 3.6.7:

    invalid status highlight color could be maliciously used to corrupt memory
    formatting corpse names used internal buffers differently from formatting
other objects and could potentially clobber memory
    during engraving, spaces were counted instead of non-space (cherry-pick of
4e0a1e04 from NetHack-3.7)
    avoid potential buffer overflow in append_str()
    resolve missing dependency in NetHack.sln
    code in include/tradstdc.h was trying to suppress warn_unused result by
defining warn_unused_result to an empty string and that began causing a build
error within a system-supplied header file cdefs.h when using Ubuntu Impish
21.10; disable that for any Linux and macOS gcc unless GCC_URWARN is defined to
force it back into effect
    update_inventory() after leash goes slack
    player assigned name for monsters, specific objects, or object types could
be longer than what was intented to be allowed; for 'curses', much longer
    curses: cherry-picked selectsaved code from 3.7 for menu of save files
    NetHackW: fix delayed rendering of cursor when using farlook