bulk-small: drop f2c.
it's been a _long_ time since pkgsrc fortran was rototilled to not
have libtool-base depending on the fortran compiler, and also a good
long time since we stopped using f2c as the fortran compiler.

I don't think f2c is really useful at all these days, but if it is
still a reasonable setup for small systems someone let me know and
we'll put it in bulk-medium.

(though it's also hard to imagine there's any significant number of
people running numeric software on small/old machines, or using
fortran for anything other than numeric software on such platforms...)