meson: fix non-default Python sandboxed builds with py-cython
Since Meson is not Python versioned, but Cython is, we previously
ended up with a situation where builds of Cython-using packages
would fail. Instead of the Makefile kludge where it would simply
force the default Python version associated with Cython, use a
different hack that will search for all (presently) possible versions.

This is not an ideal solution, but is being committed to fix build
breakage in the lead up to branching pkgsrc 2023Q3. Other solutions
would be to go back to Python versioned Meson, or getting Meson to
handle multi-Python environments more cleanly.