opa: Update to 0.57.0
Changes:
0.57.0
------
This release contains an updated Rego syntax to allow general
references in rule heads, and a mix of new features and bugfixes.
Support for General References in Rule Heads

In OPA 0.56.0, we introduced support for general references in rule
heads as an experimental feature.  It has now graduated to a fully
supported feature, and is no longer experimental.

A general reference is a reference with variables at arbitrary locations.
In Rego, partial rules are used for generating sets and objects.
In previous versions of OPA, variables were only allowed in the very
last position in the rule's reference.
Now, Rego has been expanded to allow rules to be declared with general
references in their head, with variables at arbitrary locations.