mysql57: updated to 5.7.43
Changes in MySQL 5.7.43

Compilation Notes

Improved Clang 15 usage, although it's not officially supported in v5.7.

Functionality Added or Changed

Important Change: The linked OpenSSL library for MySQL Server has been upgraded from OpenSSL 1.1.1 to OpenSSL 3.0. The exact version is now 3.0.9. More information on changes from 1.1.1 to 3.0 can be found at https://www.openssl.org/docs/man3.0/man7/migration_guide.html.

The linked curl library for MySQL Server (Enterprise Edition) has been updated to version 8.1.1.

Bugs Fixed

Group Replication: After one machine halted and restarted in a three-node MySQL InnoDB Cluster, one node failed to start; after restarting all nodes, the cluster shut down unexpectedly.

Our thanks to Zetang Zeng for the contribution.

Fortified parsing of the network packet data sent by the server to the client.

Some floating-point literals were not always handled correctly.

Executing a query with an implicit aggregation should return exactly one row, unless the query has a HAVING clause that filters out the row, but a query with a HAVING clause which evaluated to FALSE sometimes ignored this, and returned a row regardless.

During optimization, range-select tree creation uses logic which differs based on the left-hand side of the IN() predicate. For a field item, each value on the right-hand side is added to an OR tree to create the necessary expression. In the case of a row item comparison (example: WHERE (a,b) IN ((n1,m1), (n2, m2), ...)), an expression in disjunctive normal form (DNF) is needed. A DNF expression is created by adding an AND tree with column values to an OR tree for each set of RHS values, but instead the OR tree was added to the AND tree causing the tree merge to require exponential time due to O(n2) runtime complexity.