Updated security/crudesaml to 1.11
This release works around a memory management problem in libgobject that
caused the process to crash in liblasso after pam_saml.so had been
unloaded and reloaded.

The fix is to link pam_saml.so with -z nodelete so that it is never
unloaded. A configure test checks that this option does not choke
a non GNU loader, and avoids using it if it does.