coyim: update to 0.4.1, now with Go module build.
This a small point release that primarily improves the flow for in-band
registration. It removes a large number of the previously recommended servers
since they no longer support registration, and solves other issues related to
registration. This is the first release containing rudimentary support for a
global application stylesheet. This will later provide support for customized
styling by the end-user.