py-ipython: updated to 8.17.2
IPython 8.17, 8.17.1
--------------------

Medium-sized release of IPython that includes some cleanup (backcall, python2 leftovers)
and some refactoring improvements (typing, pathlib) and a fix on completion.

  - remove backcall dependency
  - make pickleshare dependency optional
  - support completion based on type annotations of calls

Reverted in 8.17.1:

  - remove support for python 2 in lexers (reverted in 8.17.1 as it is imported by qtconsole/spyder)

Mamba and Micromamba magic
~~~~~~~~~~~~~~~~~~~~~~~~~~

In addition to the conda command to manage conda environment, mamba and
micromamba can now be used using the corresponding magic in IPython.
Since these commands are compatible with conda, they are following the
same logic.

These two magic require to have the corresponding commands available
either in the conda environment or system wide.