py-pspdfutils: update to 3.3.2.
3.3.2

This release improves the documentation of pagespecs, and adds an
example

The error message for invalid pagespecs previously had a small
mistake in the syntax it displayed; this has been fixed. Thanks,
@jgclark!

3.3.1

This release fixes a bug in the pstops(1) man page example for
duplex book printing. Many apologies to users who wondered what
they were doing wrong when following it!

3.3.0

This release improves psnup and updates to a later version of pypdf
that fixes a crash.

In psnup, improve how we guess the page size. Since we can now
guess the size of PostScript files accurately in many cases, use
the input size first, then the output size, and only then fall back
to paper’s value.

Also, fix a bug where the output paper size was incorrectly changed
during processing.

We now require pypdf >= 3.16, which fixes a bug that could cause
a crash on PDF documents containing links that pointed to each
other.

Finally, the test suite has been improved to fall back to visual
comparison of PDF outputs when they are not byte-identical. This
should ease pypdf upgrades in future, and in particular avoid either
having to hard-wire a particular pypdf version (as we did with
PSUtils version 3.2.0), or risk test failures with newer versions
of pypdf than the minimum required version (as we did previously).