Add tipidee: Minimalistic web server
tipidee is a web server supporting HTTP 1.0 and 1.1. It aims to be
compliant with RFC 9112: while it only implements a very limited subset
of the optional functionality in HTTP 1.1, it implements all the
mandatory parts.

It runs under a super-server, e.g. inetd, s6-tcpserver, or s6-tlsserver
(for HTTPS). Traditionally, inetd-mode web servers aren't considered
performant, but tipidee aims to eke out every single drop of performance
that is attainable with its programming model.

Key features:

- Usability with HTTPS without the need to entangle the code with a
  given TLS library
- Support for HTTP 1.1, with persistent connections, and not only 1.0
- Support for real CGI, not only NPH