shells/nushell: update to 0.87.1
This is the 0.87.1 patch release of Nushell. In this hotfix we resolve two
issues with the file system commands rm and cp.

Nushell's cp command in 0.87.0 did not correctly handle changes to the directory
by the Nushell internal cd when invoked inside a block. Furthermore the rm
command was not able to correctly handle paths requiring expansion.