x11/py-sip6: Apply upstream fix for bug that breaks everything
The 6.8.0 was defective and caused many, if not almost all, uses of
sip to break.  E.g.  https://github.com/qgis/QGIS/issues/55481

Cherry-pick upstream change from 30 November, because for reasons I do
not understand there isn't a fixed point release.

Add a note that updates should be build tested with qgis, as
upstream's QA is clearly not good enough to assume that will be ok.