xbiff: update to 1.0.5.
This release adds -help & -version command line options.

For those building for 32-bit platforms, it also enables use of the
"large file" APIs so that it can handle mailboxes which exceed 2gb
or which may be stored on a filesystem which uses large inode numbers
not supported by the original 32-bit APIs.

Since this release was generated using the new GNU autoconf 2.72, this
also adds a --enable-year2038 configure flag which may allow a 32-bit
build of xbiff to work with files whose timestamps are later than
January 19, 2038, but this has not been tested.

Alan Coopersmith (12):
      gitlab CI: add a basic build test
      Build xz tarballs instead of bzip2
      gitlab CI: stop requiring Signed-off-by in commits
      configure: Use AC_SYS_LARGEFILE to enable large file support
      Use XtAsprintf instead of manual mallocing and copying of strings
      Combine usage message into a single string
      Add -version option
      Fix -version handling to not require opening a display first
      Add -help option
      List which options were unrecognized before printing usage message
      Use C99 struct initializers
      xbiff 1.0.5