py-jsonref: updated to 1.1.0
v1.1.0
Fixes documents where the root is a reference object not being cached
Allows a reference object to refer to a fragment within itself.
Extra properties of a reference object will now have references resolved.

v1.0.1
Fix issue with proxies=False where it would not recurse through references to remove the proxies.

v1.0.0
replace_refs is now a function rather than a class method of JsonRef (The class method is still around, but deprecated.)
New options were also added to replace_refs
jsonloader is now a plan function rather than an instance of a class. (This is a breaking change if you were subclassing JsonLoader)