digikam: update to 8.2.0.
After five months of active maintenance and long bugs triage, the
digiKam team is proud to present version 8.2.0 of its open source
digital photo manager.

See below the list of most important features coming with this release.

    Libraw : Updated to snapshot 2023-11-21

This version arrives with a long review of bugzilla entries. Long
time bugs present in older versions have been fixed and we spare
a lots of time to contact users to validate changes in pre-release
to confirm fixes before to deploy the program in production.