multimedia/ffmpeg5: Handle machdep.altivec on NetBSD/powerpc too.
Bump revision because this built without the change on
NetBSD/powerpc, but didn't respect machdep.altivec at runtime.

Don't build with -maltivec -mabi=altivec on powerpc.  This doesn't
disable the use of altivec; it just prevents the compiler from trying
to use altivec instructions, like vector-zero and vector-store to
zero-initialize an object on the stack, outside the parts that are
gated on runtime CPU feature detection.

Fixes ffmpeg5 on powerpc CPUs without altivec.