py-hypothesis: updated to 6.97.3
6.97.3

Internal test refactoring.

6.97.2

This patch slightly changes how we replay examples from the database: if the behavior of the saved example has changed, we now keep running the test case instead of aborting at the size of the saved example. While we know it’s not the same example, we might as well continue running the test!

Because we now finish running a few more examples for affected tests, this might be a slight slowdown - but correspondingly more likely to find a bug.

We’ve also applied similar tricks to the target phase, where they are a pure performance improvement for affected tests.

6.97.1

Improves the performance of the arrays() strategy when generating unique values.