tex-polyglossia{,-doc}: update to 2.0
We have bumped the version to 2.0 with this release, not so much since this is
a revolutionary issue (it isn't), but (1.) to indicate the slow (and
unfinished) transformation of the polyglossia code to latex3 which quite
progressed in this release, and (2.) simply since the minor version number got
ridiculously large meanwhile.

The most important fix in this release concerns a problem that appeared with
the recent l3kernel update and which broke compilation of many documents. As it
is quite likely that this hits, or will hit, you if you use polyglossia, we
encourage all users to upgrade.