encodings: update to 1.1.0.
This release adds the ISO8859-14 encoding

This allows legacy applications (xcalc for instance) to work without the
 "Warning: missing charsets in String to FontSet conversion"
message on systems with only ISO8859-1 and UTF-8 encoded fonts installed.

This release also adds support for building with meson as well as autoconf.
Please test building with meson and report any issues you hit to our
gitlab issue tracker at:
  https://gitlab.freedesktop.org/xorg/font/encodings/-/issues
as we plan on removing the autoconf build support in a future release.

Alan Coopersmith (2):
      Add a meson build system
      font-encodings 1.1.0

Matthieu Herrb (1):
      Add the ISO8859-14 encoding