pkg_install: Add sanity checks for pkgdb correctness.
A number of users have hit "Conflicting PLIST" errors in the past that are
almost certainly caused by duplicate PKGBASE entries in the pkgdb, most likely
as a result of unpacking a bootstrap kit over the top of an existing install.

This adds a check to ensure there are no duplicate PKGBASE entries, displaying
an error message with the list of duplicates if they are found, and is called
for the "check", "rebuild", and "rebuild-tree" commands.