py-hypothesis: updated to 6.100.1
6.100.1

This patch improve a rare error message for flaky tests.

6.100.0

The from_dtype() function no longer generates NaT (“not-a-time”) values for the datetime64 or timedelta64 dtypes if passed allow_nan=False.

6.99.13

This patch includes the backend setting in the how_generated field of our observability output.

6.99.12

If you were running Python 3.13 (currently in alpha) with pytest-xdist and then attempted to pretty-print a lambda functions which was created using the eval() builtin, it would have raised an AssertionError. Now you’ll get "lambda ...: <unknown>", as expected.

6.99.11

This release improves an internal invariant.

6.99.10

This patch fixes Hypothesis sometimes raising a Flaky error when generating collections of unique floats containing nan.