mk: Correctly engineer previous change.
Rather than breaking many of the 23 platforms that pkgsrc purports to support,
move the EXTRACT_OPTS_TAR logic into the individual package that needed it.
If there are other packages that need this they can easily copy the required
two lines.

It's worth noting here that only bsdtar has a problem with this, gtar extracts
the file fine (with warnings).