mail/mu: Add mu version 1.12.3
mu is a tool for dealing with e-mail messages stored in the
Maildir-format.  mu's purpose in life is to help you to quickly
find the messages you need; in addition, it allows you to view
messages, extract attachments, create new maildirs, and so on.

Built on top of mu are some extensions:
- mu4e: a full-featured e-mail client that runs inside emacs
- mu-guile: bindings for the Guile/Scheme programming language