rust: move pkgsrc's generated cargo config to $WRKDIR
Some setuptools-rust packages (at least py-bcrypt) won't find it in the
source directory, leading to the pkgsrc crate vendoring not having the
desired effect. Move it to ${WRKDIR}/.cargo/config.toml instead and
set CARGO_HOME in the environment such that cargo will find it regardless
of which directory it is started from.