Updated www/dillo to 3.1.0
Since the last release 3.0.5 from 2015 a lot of things have happened
to the Dillo project. A plot of number of commits over time colored
by authors. There is a missing gap from mid 2002 to mid 2008. And
from 2017 to end of 2023 there is almost no commits. Here is a
short timeline to put things into perspective:

 - On 2016, the main developer of the layout engine, Sebastian
   Geerken, passed away
 - On 2017, the development stopped
 - On 2019, the last email from Jorge Arellano Cid, the lead
   developer of Dillo, was recorded by the mailing list
 - On 2022, the domain dillo.org was lost
 - On 2024, an attempt to resurrect the project began by Rodrigo
   Arias Mallo

This release contains a lot of changes accumulated from the 2015-2017
period, as well as fixes and small features introduced in 2024.
Here is a summary:

 - Add support for floating HTML elements, which involved a big
   redesign
 - Add support for OpenSSL, LibreSSL and mbed TLS for HTTPS, which
   is now enabled by default
 - Add a CI pipeline to build Dillo on Ubuntu, MacOS, FreeBSD and
   Windows (via cygwin)
 - Add automatic HTML rendering tests
 - Improve and extend the Dillo manual