databases/mariadb1011-embedded: import mariadb-embedded-10.11.7
MariaDB Server is one of the most popular open source relational databases.
It's made by the original developers of MySQL.

MariaDB turns data into structured information in a wide array of applications,
ranging from banking to websites. It is an enhanced, drop-in replacement
for MySQL. MariaDB is used because it is fast, scalable and robust, with a
rich ecosystem of storage engines, plugins and many other tools make it very
versatile for a wide variety of use cases.

MariaDB is developed as open source software and as a relational database
it provides an SQL interface for accessing data. The latest versions of
MariaDB also include GIS and JSON features.

This package contains the embedded MariaDB server programs.