gcc.mk: Improve comments about several USE_*_FEATURES
Add link to gcc stdc++ status.

Rototill c17 comments, noting that I can't find where gcc documents
this.  Note that gcc7 fails and gcc10 works, and 8/9 are unknown.

Note that std::filesystem is part of C++17, that it might work under 9
and that 10 is due to the limited versions guideline.

Note that parallism_ts is part of C++17, that gcc docs imply 9, and
that 10 is due to the limited versions guidline.

Note that charconv is part of C++17, and that it works in gcc8.
Rewrite my earlier comment into now-standard form about limited
versions guidance.

This is comment-only change.