py-hypothesis: updated to 6.103.1
6.103.1

Account for time spent in garbage collection during tests, to avoid flaky DeadlineExceeded errors.

Also fixes overcounting of stateful run times, a minor observability bug dating to version 6.98.9.

6.103.0

This release migrates the shrinker to our new internal representation, called the IR layer. This improves the shrinker’s performance in the majority of cases. For example, on the Hypothesis test suite, shrinking is a median of 1.38x faster.

It is possible this release regresses performance while shrinking certain strategies. If you encounter strategies which reliably shrink more slowly than they used to (or shrink slowly at all), please open an issue!

6.102.6

This patch fixes one of our shrinking passes getting into a rare O(n) case instead of O(log(n)).

6.102.5

This patch fixes some introspection errors new in Python 3.11.9 and 3.13.0b1, for the Ghostwriter and from_type().

6.102.4

Internal developer documentation, no user-visible changes.

6.102.3

This patch improves our shrinking of unique collections, such as dictionaries(), sets(), and lists() with unique=True.

6.102.2

This patch fixes a rare internal error when generating very large elements from strategies.

6.102.1

This patch fixes an overly strict internal type assertion.

6.102.0

This release improves our support for the annotated-types iterable GroupedMetadata protocol. In order to treat the elements “as if they had been unpacked”, if one such element is a SearchStrategy we now resolve to that strategy. Previously, we treated this as an unknown filter predicate.

We expect this to be useful for libraries implementing custom metadata - instead of requiring downstream integration, they can implement the protocol and yield a lazily-created strategy. Doing so only if Hypothesis is in sys.modules gives powerful integration with no runtime overhead or extra dependencies.