geography/gdal*: Cope with deficient (no rtree) sqlite3
As background, the sqlite3 in NetBSD 9 lacks rtree support.

gdal expects the rtree extension to sqlite3 (geopackages are an
important format, and based on sqlite).  The base package had a cmake
configure arg to allow accepting sqlite3 without rtree.  This commit
rototills gdal-lib's comments about base/pkg sqlite3 and rtree, but
does not otherwise change gdal-lib.

In py-gdal, declare a cmake configure arg to accept sqlite3 without
rtree, and pass it to the hand-written cmake run to create bits needed
by the python build (that could have been in the distfile but aren't).
No change is expected if the package built before.

(It remains to understand if gdal on NetBSD 9 has operational failures
due to missing rtree support.)