Add protobuf-go version 1.34.1
Go support for Protocol Buffers

This project hosts the Go implementation for
[protocol buffers](https://protobuf.dev), which is a
language-neutral, platform-neutral, extensible mechanism for serializing
structured data. The protocol buffer language is a language for specifying the
schema for structured data. This schema is compiled into language specific
bindings. This project provides both a tool to generate Go code for the
protocol buffer language, and also the runtime implementation to handle
serialization of messages in Go. See the
[protocol buffer developer guide](https://protobuf.dev/overview)
for more information about protocol buffers themselves.