graphics/py-DPainter: import py-DPainter-1.2.2
PyDPainter, pronounced "Pied Painter" (like Pied Piper), is an
attempt to create a usable pixel art program in Python using PyGame.

The original inspiration came from the Commodore Amiga version of
Deluxe Paint released by Electronic Arts in 1985.  Back then, Deluxe
Paint helped define the user interface of a paint program with tool
bars, menus, and the novel use of left and right mouse buttons for
painting and erasing. After pixel art gave way to photo-realism
and high-resolution 24 bit color, Deluxe Paint was largely forgotten
for artistic work -- left behind in the ever-progressing march of
technology.

Recently, with a resurgence of all things "retro," low-resolution
pixel art and limited color palettes have become popular once again.
Many tools to deal with this medium are either too complicated or
too crude. This project is an attempt to bring back an old but
reliable tool and enhance it with some features to help it better
coexist in the modern world.