databases/mongodb3: Remediate icu, fixing build on NetBSD 9
Declare that a C++11 compiler is needed.  (This has surely been true
for a very long time, but wasn't in the Makefile.)

Recently icu apparently grew a requirement that files including icu
headers be compiled as C++17, and this package was therefore marked to
ask for C++17.  That apparently worked on NetBSD 10, but on NetBSD 9,
it failed to build.   Resolve this by removing the C++17 FORCE_CXX_STD
(which makes no sense for mongodb3 which needs C++11), and use the
included icu, taking a hint from how included boost is used.

(This is a pullup candidate for 2024Q2, should anyone else care and
want to do that.)