py-debugpy: updated to 1.8.5
debugpy v1.8.5

To simplify command line usage, an entry point has been added to the debugpy package.

If you install debugpy into an environment via pip, you can just run debugpy from that environment, instead of running python -m debugpy
If you install debugpy into an isolated directory, the debugpy executable can be found in the bin directory inside of the install directory. In this case, you need to add the install directory to your PYTHONPATH in order for the imports to work.
To simply command line usage, switches can now be read from the environment.

To do this, use the variable DEBUGPY_EXTRA_ARGV and add whatever command-line switches you like.
For example, set DEBUGPY_EXTRA_ARGV=--connect 5678 will allow you to omit the --connect switch from the command line.
If the same switch is found on both the command line and the environment, the one from the environment is ignored.
The switch that specifies the "target" MUST be on the command line (filename, -m <module>, -c <code>, or --pid <id>)