Update to 1.43
Upstream changes:
1.43 Nov 7 2021
Our implementation of "which" should not search for a program in the elements of PATH when program already contains a path. [rjung]

Fix hash generation when the openssl command was overwritten with the environment variable "APACHE_TEST_OPENSSL_CMD". [rjung]

Add new config option limitrequestline needed when the local directory paths are long. [rjung]