net/py-foolscap: Update to 23.11.0
Use github, and kludge github.mk for the distfile format, because pypi
lacks a source archive.

Use WRKSRC for the bizarre unpack location of the github source archive.

Upstream lacks NEWS but github release notes indicate that changes are
only about adapting to newer python versions (and dropping 3.7 which
is ancient).