palemoon: Import the rest of our nsprpub patches
It's unclear to me if these are strictly necessary with Pale Moon.
They appear to not affect browser stability for me.  However, since
they're bugs in the nsprpub library that Mozilla doesn't want to fix
(likely due to API compatibility?), and previously triggered crashes on
NetBSD, better safe than sorry.