sysutils/py-borgbackup: make this accept a newer py-msgpack.
This is in contravention to the upstream policy which insists that
because the maintainers of py-msgpack in the distant past made a
non-compatible change (apparently in a version before 0.5.6 of
py-msgpack, many, many years ago), they now insist that you *cannot*
use a newer version of py-msgpack than what the code insists on,
ref.

   https://github.com/borgbackup/borg/issues/3753

This collides (hard) with the pkgsrc default stance which is to
"upgrade everything to the newest available version", and makes
this package break each time a new version of py-msgpack is integrated
in pkgsrc.

For now just patch this to accept py-msgpack versions between
0.5.6 and 1.1.0 and not just between 0.5.6 and 1.0.8.

At least testing by doing a backup run makes that backup run
complete successfully with py-mgspack 1.1.0 installed, instead
of a sour error message that a non-supported version of msgpack
is installed.